@@ -5,15 +5,15 @@ abstract class SessionStatusInfo extends JSONAble {
   SessionStatusInfo(this.loggedIn, this.timestamp, {this.currentUser});
 
   /// Whether the user is logged in or not.
-  bool loggedIn;
+  bool? loggedIn;
 
   /// The timestamp of the session when updated.
-  double timestamp;
+  double? timestamp;
 
   /// The current user logged in.
-  String currentUser;
+  String? currentUser;
 
   /// Holds the complete JSON retrieved from the backend
   /// This is because each app will have custom fields in the session response.
-  Map<String, dynamic> rawSession;
+  Map<String, dynamic>? rawSession;
 }
