@@ -9,8 +9,8 @@ part 'fcm_notification.model.g.dart';
 class FCMNotification extends JSONAble {
   FCMNotification();
 
-  factory FCMNotification.fromJson(Map<String, dynamic> json) {
-    json['notification'] = json['notification']?.cast<String, dynamic>();
+  factory FCMNotification.fromJson(Map<String, dynamic>? json) {
+    json!['notification'] = json['notification']?.cast<String, dynamic>();
     json['data'] = json['data'] != null
         ? json['data'].cast<String, dynamic>()
         // Handle iOS Notifications where it doesn't contain data object (All are in one Map)
@@ -18,32 +18,32 @@ class FCMNotification extends JSONAble {
     return _$FCMNotificationFromJson(json);
   }
 
-  NotificationMessage notification;
+  NotificationMessage? notification;
 
-  Map<String, dynamic> data;
+  Map<String, dynamic>? data;
 
-  String title;
-  String body;
+  String? title;
+  String? body;
 
   @JsonKey(name: 'message_id')
-  String messageId;
+  String? messageId;
 
   @JsonKey(
       fromJson: FrappeDocFieldConverter.checkToBool,
       toJson: FrappeDocFieldConverter.boolToCheck)
-  bool seen;
+  bool? seen;
 
   @JsonKey(name: 'communication_date')
-  String communicationDate;
+  String? communicationDate;
 
   // Useful for iOS where it appends the fields within data and omits data key
-  Map<String, dynamic> rawNotification;
+  Map<String, dynamic>? rawNotification;
 
   @override
   Map<String, dynamic> toJson() => _$FCMNotificationToJson(this);
 
   @override
-  T fromJson<T>(Map<String, dynamic> json) =>
+  T fromJson<T>(Map<String, dynamic>? json) =>
       FCMNotification.fromJson(json) as T;
 }
 
@@ -54,16 +54,16 @@ class FCMNotification extends JSONAble {
 class NotificationMessage extends JSONAble {
   NotificationMessage();
 
-  factory NotificationMessage.fromJson(Map<String, dynamic> json) =>
-      _$NotificationMessageFromJson(json);
+  factory NotificationMessage.fromJson(Map<String, dynamic>? json) =>
+      _$NotificationMessageFromJson(json!);
 
-  String title;
-  String body;
+  String? title;
+  String? body;
 
   @override
   Map<String, dynamic> toJson() => _$NotificationMessageToJson(this);
 
   @override
-  T fromJson<T>(Map<String, dynamic> json) =>
+  T fromJson<T>(Map<String, dynamic>? json) =>
       NotificationMessage.fromJson(json) as T;
 }
