@@ -36,7 +36,7 @@ enum RenovationError {
 }
 
 class Request {
-  static String clientId;
+  static String? clientId;
 
   /// The [Dio] instance used in all network calls
   static final Dio _dio = Dio();
@@ -44,19 +44,19 @@ class Request {
   /// Persistent cookie when JWT is not in use.
   ///
   /// Initialized when [setupPersistentCookie] is called
-  static PersistCookieJar _cookieJar;
+  static PersistCookieJar? _cookieJar;
 
   /// Handles the raw request to a server
   ///
   /// Supported methods are *GET*, *POST*, *PUT*, *DELETE*,
   ///
   static Future<Response<String>> _httpRequest(
-      String url, HttpMethod httpMethod, Map<String, dynamic> headers,
-      {String contentType,
-      Map<String, dynamic> data,
-      Map<String, dynamic> queryParams}) async {
+      String? url, HttpMethod httpMethod, Map<String, dynamic>? headers,
+      {String? contentType,
+      Map<String, dynamic>? data,
+      Map<String, dynamic>? queryParams}) async {
     final httpOptions = Options();
-    ContentType _contentType;
+    late ContentType _contentType;
 
     headers ??= RenovationRequestOptions.headers;
 
@@ -81,23 +81,23 @@ class Request {
     }
 
     if (clientId != null) {
-      headers['X-Client-Site'] = clientId;
+      headers!['X-Client-Site'] = clientId;
     }
     httpOptions.headers = headers;
     httpOptions.contentType = _contentType.value;
 
     switch (httpMethod) {
       case HttpMethod.GET:
-        return _dio.get<String>(url,
+        return _dio.get<String>(url!,
             options: httpOptions, queryParameters: queryParams);
       case HttpMethod.POST:
-        return _dio.post<String>(url,
+        return _dio.post<String>(url!,
             options: httpOptions, data: data, queryParameters: queryParams);
       case HttpMethod.PUT:
-        return _dio.put<String>(url,
+        return _dio.put<String>(url!,
             options: httpOptions, data: data, queryParameters: queryParams);
       case HttpMethod.DELETE:
-        return _dio.delete<String>(url,
+        return _dio.delete<String>(url!,
             options: httpOptions, queryParameters: queryParams);
         break;
       default:
@@ -107,23 +107,23 @@ class Request {
     }
   }
 
-  static Future<RequestResponse<FrappeResponse>> initiateRequest(
-      {@required String url,
-      @required HttpMethod method,
-      String contentType,
-      Map<String, dynamic> headers,
-      Map<String, dynamic> data,
-      Map<String, dynamic> queryParams,
+  static Future<RequestResponse<FrappeResponse?>> initiateRequest(
+      {required String? url,
+      required HttpMethod method,
+      String? contentType,
+      Map<String, dynamic>? headers,
+      Map<String, dynamic>? data,
+      Map<String, dynamic>? queryParams,
       bool isFrappeResponse = true}) async {
-    Response<dynamic> response;
-    RequestResponse<FrappeResponse> r;
+    Response<dynamic>? response;
+    RequestResponse<FrappeResponse?>? r;
     try {
       response = await Request._httpRequest(url, method, headers,
           contentType: contentType, data: data, queryParams: queryParams);
 
       final frappeResponse = _buildFrappeResponse(response, isFrappeResponse);
 
-      if ((response.statusCode / 100).floor() == 2) {
+      if ((response.statusCode! / 100).floor() == 2) {
         r = RequestResponse.success<FrappeResponse>(frappeResponse);
       } else {
         final info =
@@ -134,7 +134,7 @@ class Request {
     } on DioError catch (err) {
       if (err.response != null) {
         final frappeResponse =
-            _buildFrappeResponse(err.response, isFrappeResponse);
+            _buildFrappeResponse(err.response!, isFrappeResponse);
         final info = Information(
             httpCode: err.response?.statusCode,
             data: frappeResponse,
@@ -157,21 +157,21 @@ class Request {
           headers: headers);
     }
 
-    r.rawResponse = response;
+    r.rawResponse = response as Response<String>?;
     return r;
   }
 
-  static void setClientId(String id) {
+  static void setClientId(String? id) {
     clientId = id;
-    RenovationConfig.renovationInstance.logger.i('Client ID: $clientId');
+    RenovationConfig.renovationInstance!.logger.i('Client ID: $clientId');
   }
 
-  static String getClientId() => clientId;
+  static String? getClientId() => clientId;
 
   static FrappeResponse _buildFrappeResponse(
       Response<dynamic> response, bool isFrappeResponse) {
-    Map<String, dynamic> jsonResponse;
-    final isJsonResponse = response.headers[Headers.contentTypeHeader][0] ==
+    Map<String, dynamic>? jsonResponse;
+    final isJsonResponse = response.headers[Headers.contentTypeHeader]![0] ==
         ContentType.json.value;
     if (isJsonResponse) {
       jsonResponse = json.decode(response.data);
@@ -182,7 +182,7 @@ class Request {
         frappeResponse = FrappeResponse.fromJson(jsonResponse);
       } else {
         frappeResponse
-          ..message = jsonResponse['data'] ?? jsonResponse
+          ..message = jsonResponse!['data'] ?? jsonResponse
           ..serverMessages = jsonResponse['server_messages']
           ..exc = jsonResponse['exc']
           ..excType = jsonResponse['exc_type']
@@ -195,41 +195,41 @@ class Request {
     return frappeResponse;
   }
 
-  static void _messageChecker(RequestResponse<FrappeResponse> response) {
+  static void _messageChecker(RequestResponse<FrappeResponse?> response) {
     final Function addMessage = (dynamic message) {
       if (message is String) {
         message = json.decode(message);
       }
-      RenovationConfig.renovationInstance.coreInstance.messages.add(message);
+      RenovationConfig.renovationInstance!.coreInstance.messages.add(message);
     };
 
-    List<dynamic> serverMessagesArray;
-    if (response.isSuccess && response.data.serverMessages != null) {
-      serverMessagesArray = json.decode(response.data.serverMessages);
+    List<dynamic>? serverMessagesArray;
+    if (response.isSuccess && response.data!.serverMessages != null) {
+      serverMessagesArray = json.decode(response.data!.serverMessages!);
 
-      serverMessagesArray.forEach(addMessage);
+      serverMessagesArray!.forEach(addMessage as void Function(dynamic));
     }
     if ((!response.isSuccess &&
-        response.error.info.data is FrappeResponse &&
-        response.error.info.data.serverMessages != null)) {
+        response.error!.info!.data is FrappeResponse &&
+        response.error!.info!.data.serverMessages != null)) {
       serverMessagesArray = List<String>.from(
-          json.decode(response.error.info.data.serverMessages));
+          json.decode(response.error!.info!.data.serverMessages));
 
-      serverMessagesArray.forEach(addMessage);
+      serverMessagesArray.forEach(addMessage as void Function(dynamic));
     }
   }
 
-  static void _errorChecker(RequestResponse<FrappeResponse> r) {
+  static void _errorChecker(RequestResponse<FrappeResponse?> r) {
     if (!r.isSuccess) {
       final Function excContains = (String sx) {
-        if (r.error.info.data.exc != null) {
-          return r.error.info.data.exc.contains(sx);
+        if (r.error!.info!.data.exc != null) {
+          return r.error!.info!.data.exc.contains(sx);
         }
         return false;
       };
       if (r.data != null) {
-        if (r.data.sessionExpired == null || r.data.sessionExpired == 1) {
-          Renovation().auth.resetSession();
+        if (r.data!.sessionExpired == null || r.data!.sessionExpired == 1) {
+          Renovation().auth!.resetSession();
         }
       }
 
@@ -237,30 +237,30 @@ class Request {
 
       if (excContains('PermissionError')) {
         r.exc = RenovationError.PermissionError;
-        r.error.type = RenovationError.PermissionError;
+        r.error!.type = RenovationError.PermissionError;
       } else if (excContains('AuthenticationError')) {
         r.exc = RenovationError.AuthenticationError;
-        r.error.type = RenovationError.AuthenticationError;
+        r.error!.type = RenovationError.AuthenticationError;
       }
-      if (r.error.info.data.serverMessages != null) {
+      if (r.error!.info!.data.serverMessages != null) {
         // Array of server message strings that still need to be decoded
         final serverMessagesArray =
-            List<String>.from(json.decode(r.error.info.data.serverMessages));
+            List<String>.from(json.decode(r.error!.info!.data.serverMessages));
 
         final serverMessagesArrayParsed = serverMessagesArray
-            .map<Map<String, dynamic>>((String message) => json.decode(message))
+            .map<Map<String, dynamic>?>((String message) => json.decode(message))
             .toList();
         if (serverMessagesArrayParsed.isNotEmpty) {
-          if (serverMessagesArrayParsed[0].containsKey('message')) {
-            if (serverMessagesArrayParsed[0]['message'] == 'Invalid Request') {
-              RenovationConfig.renovationInstance.logger
+          if (serverMessagesArrayParsed[0]!.containsKey('message')) {
+            if (serverMessagesArrayParsed[0]!['message'] == 'Invalid Request') {
+              RenovationConfig.renovationInstance!.logger
                   .e('Invalid Request. Best to re-login');
               try {
-                RenovationConfig.renovationInstance.coreInstance.auth.logout();
+                RenovationConfig.renovationInstance!.coreInstance.auth!.logout();
               } on Exception catch (e) {
-                RenovationConfig.renovationInstance.logger.e(e);
+                RenovationConfig.renovationInstance!.logger.e(e);
               }
-              Renovation().auth.resetSession();
+              Renovation().auth!.resetSession();
             }
           }
         }
@@ -269,7 +269,7 @@ class Request {
   }
 
   /// Returns the JSON as a [Map] from the response, if any.
-  static Map<String, dynamic> convertToMap(Response<dynamic> response) {
+  static Map<String, dynamic>? convertToMap(Response<dynamic> response) {
     if (response.data != null && isJsonResponse(response)) {
       return json.decode(response.data);
     }
@@ -278,15 +278,15 @@ class Request {
 
   /// Returns `true` if the [Response] or [DioError] contains a JSON response.
   static bool isJsonResponse(Response<dynamic> response) =>
-      response.headers[Headers.contentTypeHeader][0] ==
+      response.headers[Headers.contentTypeHeader]![0] ==
       ContentTypeLiterals.APPLICATION_JSON;
 
   /// Set [PersistCookieJar] with the directory path.
   static void setupPersistentCookie(String dir) {
     // If the interceptor wasn't added to Dio
     if (_dio.interceptors.whereType<CookieManager>().isEmpty) {
-      _cookieJar = PersistCookieJar(dir: dir);
-      _dio.interceptors.add(CookieManager(_cookieJar));
+      // _cookieJar = PersistCookieJar(dir: dir);
+      // _dio.interceptors.add(CookieManager(_cookieJar));
     }
   }
 }
@@ -300,9 +300,9 @@ class RequestException implements Exception {
       this.headers,
       this.contentType});
 
-  String cause;
-  String url;
-  String contentType;
+  String? cause;
+  String? url;
+  String? contentType;
   dynamic data;
   dynamic queryParams;
   dynamic headers;
@@ -314,7 +314,7 @@ class RequestException implements Exception {
 
 /// Static class holding the headers to be used by each request.
 class RenovationRequestOptions {
-  static Map<String, dynamic> headers;
+  static Map<String, dynamic>? headers;
 }
 
 enum HttpMethod { GET, POST, PUT, DELETE }
