@@ -20,7 +20,7 @@ import 'utils.dart';
 
 /// A controller containing methods and properties related to document CRUD operations of Frappé
 class FrappeModelController extends ModelController<FrappeDocument> {
-  FrappeModelController(RenovationConfig config) : super(config);
+  FrappeModelController(RenovationConfig? config) : super(config);
 
   static const String DOCTYPE_NOT_EXIST_TITLE = "DocType doesn't exist";
   static const String DOCNAME_NOT_EXIST_TITLE = "Docname doesn't exist";
@@ -37,7 +37,7 @@ class FrappeModelController extends ModelController<FrappeDocument> {
     doc.isLocal = true;
     doc.unsaved = true;
     doc.rawResponse = <String, dynamic>{
-      if (doc.rawResponse != null) ...doc.rawResponse,
+      if (doc.rawResponse != null) ...doc.rawResponse!,
       ...doc.toJson(),
     };
     addToLocals(doc);
@@ -49,13 +49,17 @@ class FrappeModelController extends ModelController<FrappeDocument> {
   /// Depends on the document count saved in [newNameCount].
   @override
   @protected
-  String getNewName(String doctype) {
+  String getNewName(String? doctype) {
     EmptyDoctypeError.verify(doctype);
 
     if (ModelController.newNameCount[doctype] == null) {
       ModelController.newNameCount[doctype] = 0;
     }
-    ModelController.newNameCount[doctype]++;
+    if (ModelController.newNameCount[doctype] != null) {
+      ModelController.newNameCount[doctype] =
+          ModelController.newNameCount[doctype]! + 1;
+    }
+    ;
     return 'New $doctype ${ModelController.newNameCount[doctype]}';
   }
 
@@ -75,11 +79,11 @@ class FrappeModelController extends ModelController<FrappeDocument> {
   /// ```
   ///
   @override
-  void addToLocals<T extends FrappeDocument>(T doc) {
-    EmptyDoctypeError.verify(doc.doctype);
+  void addToLocals<T extends FrappeDocument?>(T doc) {
+    EmptyDoctypeError.verify(doc!.doctype);
 
     if (!locals.containsKey(doc.doctype)) {
-      locals[doc.doctype] = <String, dynamic>{};
+      locals[doc.doctype] = <String?, dynamic>{};
     }
     if (doc.name == null) {
       doc.name = getNewName(doc.doctype);
@@ -87,7 +91,7 @@ class FrappeModelController extends ModelController<FrappeDocument> {
       doc.isLocal = true;
       doc.unsaved = true;
     }
-    locals[doc.doctype][doc.name] = doc;
+    locals[doc.doctype]![doc.name] = doc;
   }
 
   /// Returns the document specified by its [docName].
@@ -100,10 +104,10 @@ class FrappeModelController extends ModelController<FrappeDocument> {
   ///
   /// Throws [EmptyDoctypeError] if the doctype field is not set within the model class.
   @override
-  Future<RequestResponse<T>> getDoc<T extends FrappeDocument>(
-      T obj, String docName,
-      {bool forceFetch = false}) async {
-    await getFrappe()
+  Future<RequestResponse<T?>> getDoc<T extends FrappeDocument>(
+      T obj, String? docName,
+      {bool? forceFetch = false}) async {
+    await getFrappe()!
         .checkAppInstalled(features: ['getDoc'], throwError: false);
 
     EmptyDoctypeError.verify(obj.doctype);
@@ -111,24 +115,24 @@ class FrappeModelController extends ModelController<FrappeDocument> {
 
     final cachedDoc = getDocFromCache<T>(obj.doctype, docName);
 
-    if (cachedDoc != null && !forceFetch) {
+    if (cachedDoc != null && !forceFetch!) {
       return RequestResponse.success(cachedDoc);
     }
 
-    RequestResponse<FrappeResponse> response;
+    RequestResponse<FrappeResponse?> response;
     // Default URL
     var url =
-        '${config.hostUrl}/api/resource/${Uri.encodeComponent(obj.doctype)}/${Uri.encodeComponent(docName)}';
-    if (getFrappe().getAppsVersion('renovation_core') != null) {
-      unawaited(config.coreInstance.meta.getDocMeta(doctype: obj.doctype));
+        '${config!.hostUrl}/api/resource/${Uri.encodeComponent(obj.doctype!)}/${Uri.encodeComponent(docName!)}';
+    if (getFrappe()!.getAppsVersion('renovation_core') != null) {
+      unawaited(config!.coreInstance.meta!.getDocMeta(doctype: obj.doctype));
       url =
-          '${config.hostUrl}/api/method/renovation/doc/${Uri.encodeComponent(obj.doctype)}/${Uri.encodeComponent(docName)}';
+          '${config!.hostUrl}/api/method/renovation/doc/${Uri.encodeComponent(obj.doctype!)}/${Uri.encodeComponent(docName)}';
     }
 
     response = await Request.initiateRequest(
         url: url, method: HttpMethod.GET, isFrappeResponse: false);
     if (response.isSuccess) {
-      final dynamic responseObj = response.data.message;
+      final dynamic responseObj = response.data!.message;
       if (responseObj != null && responseObj != 'failed') {
         final temp = obj.fromJson<T>(responseObj);
         temp.rawResponse = responseObj;
@@ -145,10 +149,10 @@ class FrappeModelController extends ModelController<FrappeDocument> {
                       data: response.data,
                       httpCode: response.httpCode,
                       rawResponse: response.rawResponse)
-                    ..rawError = response?.error?.info?.rawError)));
+                    ..rawError = response?.error?.info?.rawError))!);
     }
 
-    return RequestResponse.fail(handleError('get_doc', response.error));
+    return RequestResponse.fail(handleError('get_doc', response.error)!);
   }
 
   /// Returns list of documents [T] of a doctype specified by [obj].
@@ -170,16 +174,16 @@ class FrappeModelController extends ModelController<FrappeDocument> {
   ///    - {{field_name}} asc|desc
   ///    - By default, the list is ordered by modification timestamp (modified)
   @override
-  Future<RequestResponse<List<T>>> getList<T extends FrappeDocument>(T obj,
-      {List<String> fields,
+  Future<RequestResponse<List<T>?>> getList<T extends FrappeDocument>(T obj,
+      {List<String>? fields,
       dynamic filters,
-      String orderBy = 'modified desc',
-      int limitPageStart = 0,
-      int limitPageLength = 99,
-      String parent,
-      Map<String, List<String>> tableFields,
-      List<String> withLinkFields}) async {
-    await getFrappe()
+      String? orderBy = 'modified desc',
+      int? limitPageStart = 0,
+      int? limitPageLength = 99,
+      String? parent,
+      Map<String, List<String>>? tableFields,
+      List<String>? withLinkFields}) async {
+    await getFrappe()!
         .checkAppInstalled(features: ['getList'], throwError: false);
 
     orderBy ??= 'modified desc';
@@ -207,33 +211,33 @@ class FrappeModelController extends ModelController<FrappeDocument> {
     // Whether custom features are used
     final isUsingCustomFeatures = tableFields != null || withLinkFields != null;
 
-    if (getFrappe().getAppsVersion('renovation_core') == null &&
+    if (getFrappe()!.getAppsVersion('renovation_core') == null &&
         isUsingCustomFeatures) {
       // Will throw an error
-      await getFrappe()
+      await getFrappe()!
           .checkAppInstalled(features: ['tableFields', 'withLinkFields']);
     }
 
-    params.cmd = getFrappe().getAppsVersion('renovation_core') != null
+    params.cmd = getFrappe()!.getAppsVersion('renovation_core') != null
         ? 'renovation_core.db.query.get_list_with_child'
         : 'frappe.client.get_list';
 
     final response = await Request.initiateRequest(
-        url: config.hostUrl,
+        url: config!.hostUrl,
         method: HttpMethod.POST,
         contentType: ContentTypeLiterals.QUERY_STRING,
         queryParams: params.toJson());
     if (response.isSuccess) {
       // Convert the JSON => T
-      final responseObj = obj.deserializeList<T>(response.data.message);
+      final responseObj = obj.deserializeList<T>(response.data!.message)!;
       for (var i = 0; i < responseObj.length; i++) {
-        responseObj[i].rawResponse = response.data.message[i];
+        responseObj[i].rawResponse = response.data!.message[i];
       }
 
       return RequestResponse.success(responseObj,
           rawResponse: response.rawResponse);
     } else {
-      return RequestResponse.fail(handleError('get_list', response.error));
+      return RequestResponse.fail(handleError('get_list', response.error)!);
     }
   }
 
@@ -247,8 +251,8 @@ class FrappeModelController extends ModelController<FrappeDocument> {
   ///
   /// Returns a failure in case [doctype], [docName] or [docField] don't exist.
   @override
-  Future<RequestResponse<Map<String, dynamic>>> getValue(
-      String doctype, String docName, String docField) async {
+  Future<RequestResponse<Map<String, dynamic>?>> getValue(
+      String doctype, String? docName, String docField) async {
     EmptyDoctypeError.verify(doctype);
     EmptyDocNameError.verify(docName);
     EmptyDocFieldError.verify(docField);
@@ -257,13 +261,13 @@ class FrappeModelController extends ModelController<FrappeDocument> {
     params.cmd = 'frappe.client.get_value';
 
     final response = await Request.initiateRequest(
-        url: config.hostUrl,
+        url: config!.hostUrl,
         method: HttpMethod.GET,
         contentType: ContentTypeLiterals.QUERY_STRING,
         queryParams: params.toJson());
     if (response.isSuccess) {
       if (response.data?.message != null) {
-        return RequestResponse.success(response.data.message,
+        return RequestResponse.success(response.data!.message,
             rawResponse: response.rawResponse);
       } else {
         response.isSuccess = false;
@@ -273,10 +277,10 @@ class FrappeModelController extends ModelController<FrappeDocument> {
                 info: Information(
                     data: response.data,
                     httpCode: response.httpCode,
-                    rawResponse: response.rawResponse))));
+                    rawResponse: response.rawResponse)))!);
       }
     } else {
-      return RequestResponse.fail(handleError('get_value', response.error));
+      return RequestResponse.fail(handleError('get_value', response.error)!);
     }
   }
 
@@ -290,31 +294,31 @@ class FrappeModelController extends ModelController<FrappeDocument> {
   ///
   /// Returns a failure if the document or doctype don't exist.
   @override
-  Future<RequestResponse<String>> deleteDoc(
+  Future<RequestResponse<String?>> deleteDoc(
       String doctype, String docName) async {
     EmptyDoctypeError.verify(doctype);
     EmptyDocNameError.verify(docName);
 
     final response = await Request.initiateRequest(
         url:
-            '${config.hostUrl}/api/resource/${Uri.encodeComponent(doctype)}/${Uri.encodeComponent(docName)}',
+            '${config!.hostUrl}/api/resource/${Uri.encodeComponent(doctype)}/${Uri.encodeComponent(docName)}',
         method: HttpMethod.DELETE);
     if (response.isSuccess) {
-      if (locals[doctype] != null && locals[doctype][docName] != null) {
-        locals[doctype].remove(docName);
+      if (locals[doctype] != null && locals[doctype]![docName] != null) {
+        locals[doctype]!.remove(docName);
       }
       return RequestResponse.success(docName);
     } else {
-      return RequestResponse.fail(handleError('delete_doc', response.error));
+      return RequestResponse.fail(handleError('delete_doc', response.error)!);
     }
   }
 
   /// Returns a cloned doc, with a new local name.
   @override
-  T copyDoc<T extends FrappeDocument>(T doc) {
-    EmptyDoctypeError.verify(doc.doctype);
+  T copyDoc<T extends FrappeDocument?>(T doc) {
+    EmptyDoctypeError.verify(doc!.doctype);
     final clone = JSONAble.clone<T>(doc);
-    clone.name = getNewName(doc.doctype);
+    clone!.name = getNewName(doc.doctype);
     clone.docStatus = FrappeDocStatus.Draft;
     clone.isLocal = true;
     clone.unsaved = true;
@@ -324,15 +328,15 @@ class FrappeModelController extends ModelController<FrappeDocument> {
 
   /// Clones a doc, set amended_from property to the original doc.
   @override
-  T amendDoc<T extends FrappeDocument>(T doc) {
-    EmptyDoctypeError.verify(doc.doctype);
+  T amendDoc<T extends FrappeDocument?>(T doc) {
+    EmptyDoctypeError.verify(doc!.doctype);
     EmptyDocNameError.verify(doc.name);
 
     if (doc.docStatus != FrappeDocStatus.Submitted) {
       throw NotASubmittedDocument();
     }
     final newDoc = copyDoc(doc);
-    newDoc.amendedFrom = doc.name;
+    newDoc!.amendedFrom = doc.name;
     return newDoc;
   }
 
@@ -347,8 +351,8 @@ class FrappeModelController extends ModelController<FrappeDocument> {
   ///
   /// Throws [InvalidFrappeFieldValue] if [docValue] doesn't comply with [DBFilter.isDBValue].
   @override
-  Future<RequestResponse<T>> setValue<T extends FrappeDocument>(
-      T obj, String docName, String docField, dynamic docValue) async {
+  Future<RequestResponse<T?>> setValue<T extends FrappeDocument>(
+      T obj, String? docName, String docField, dynamic docValue) async {
     EmptyDoctypeError.verify(obj.doctype);
     EmptyDocNameError.verify(docName);
     EmptyDocFieldError.verify(docField);
@@ -359,16 +363,16 @@ class FrappeModelController extends ModelController<FrappeDocument> {
     params.cmd = 'frappe.client.set_value';
 
     final response = await Request.initiateRequest(
-        url: config.hostUrl,
+        url: config!.hostUrl,
         method: HttpMethod.POST,
         contentType: ContentTypeLiterals.APPLICATION_JSON,
         data: params.toJson());
     if (response.isSuccess) {
-      final temp = obj.fromJson<T>(response.data.message);
-      temp.rawResponse = response.data.message;
+      final temp = obj.fromJson<T>(response.data!.message);
+      temp.rawResponse = response.data!.message;
       return RequestResponse.success(temp, rawResponse: response.rawResponse);
     }
-    return RequestResponse.fail(handleError('set_value', response.error));
+    return RequestResponse.fail(handleError('set_value', response.error)!);
   }
 
   /// Returns the saved document [T] after saving in the backend and locally.
@@ -377,21 +381,21 @@ class FrappeModelController extends ModelController<FrappeDocument> {
   ///
   /// If a new doc is saved with the same name, a failure is returned.
   @override
-  Future<RequestResponse<T>> saveDoc<T extends FrappeDocument>(T doc) async {
-    await getFrappe()
+  Future<RequestResponse<T?>> saveDoc<T extends FrappeDocument>(T doc) async {
+    await getFrappe()!
         .checkAppInstalled(features: ['saveDoc'], throwError: false);
 
     EmptyDoctypeError.verify(doc.doctype);
     EmptyDocNameError.verify(doc.name);
 
     final isRenovationCoreInstalled =
-        getFrappe().getAppsVersion('renovation_core') != null;
+        getFrappe()!.getAppsVersion('renovation_core') != null;
 
     final response = await Request.initiateRequest(
         url: isRenovationCoreInstalled
-            ? '${config.hostUrl}/api/method/renovation/doc/${Uri.encodeComponent(doc.doctype)}/${doc.isLocal ? "" : '${Uri.encodeComponent(doc.name)}'}'
-            : '${config.hostUrl}/api/resource/${Uri.encodeComponent(doc.doctype)}${doc.isLocal ? "" : '/${Uri.encodeComponent(doc.name)}'}',
-        method: doc.isLocal ? HttpMethod.POST : HttpMethod.PUT,
+            ? '${config!.hostUrl}/api/method/renovation/doc/${Uri.encodeComponent(doc.doctype!)}/${doc.isLocal! ? "" : '${Uri.encodeComponent(doc.name!)}'}'
+            : '${config!.hostUrl}/api/resource/${Uri.encodeComponent(doc.doctype!)}${doc.isLocal! ? "" : '/${Uri.encodeComponent(doc.name!)}'}',
+        method: doc.isLocal! ? HttpMethod.POST : HttpMethod.PUT,
         contentType: ContentTypeLiterals.APPLICATION_JSON,
         data: isRenovationCoreInstalled
             ? <String, dynamic>{'doc': doc.toJson()}
@@ -399,11 +403,11 @@ class FrappeModelController extends ModelController<FrappeDocument> {
         isFrappeResponse: false);
 
     if (response.isSuccess) {
-      if (response.data != null && response.data.message is Map) {
-        final savedDoc = doc.fromJson<T>(response.data.message);
+      if (response.data != null && response.data!.message is Map) {
+        final savedDoc = doc.fromJson<T>(response.data!.message);
         savedDoc.isLocal = false;
         savedDoc.unsaved = false;
-        savedDoc.rawResponse = response.data.message;
+        savedDoc.rawResponse = response.data!.message;
         addToLocals(savedDoc);
         return RequestResponse.success(savedDoc,
             rawResponse: response.rawResponse);
@@ -417,7 +421,7 @@ class FrappeModelController extends ModelController<FrappeDocument> {
                 info: Information(
                     data: response.data,
                     rawResponse: response.rawResponse,
-                    httpCode: response.httpCode))));
+                    httpCode: response.httpCode)))!);
   }
 
   /// Submit a submittable document.
@@ -426,19 +430,19 @@ class FrappeModelController extends ModelController<FrappeDocument> {
   ///
   /// Throws [NotSubmittableDocError] if the doctype is not submittable based on its meta.
   @override
-  Future<RequestResponse<T>> submitDoc<T extends FrappeDocument>(T doc) async {
+  Future<RequestResponse<T?>> submitDoc<T extends FrappeDocument>(T doc) async {
     EmptyDoctypeError.verify(doc.doctype);
     EmptyDocNameError.verify(doc.name);
 
     final meta =
-        await getFrappeMetaController().getDocMeta(doctype: doc.doctype);
+        await getFrappeMetaController()!.getDocMeta(doctype: doc.doctype);
 
-    if (meta.isSuccess && !meta.data.isSubmittable) {
+    if (meta.isSuccess && !meta.data!.isSubmittable!) {
       throw NotSubmittableDocError();
     }
 
     final response = await Request.initiateRequest(
-        url: config.hostUrl,
+        url: config!.hostUrl,
         method: HttpMethod.POST,
         contentType: ContentTypeLiterals.APPLICATION_JSON,
         data: <String, dynamic>{
@@ -447,12 +451,12 @@ class FrappeModelController extends ModelController<FrappeDocument> {
         });
     if (response.isSuccess) {
       if (response.data != null &&
-          response.data.exc == null &&
-          response.data.message is Map) {
-        doc = doc.fromJson(response.data.message);
+          response.data!.exc == null &&
+          response.data!.message is Map) {
+        doc = doc.fromJson(response.data!.message);
         doc.isLocal = false;
         doc.unsaved = false;
-        doc.rawResponse = response.data.message;
+        doc.rawResponse = response.data!.message;
 
         addToLocals(doc);
         return RequestResponse.success(doc, rawResponse: response.rawResponse);
@@ -466,7 +470,7 @@ class FrappeModelController extends ModelController<FrappeDocument> {
                 info: Information(
                     data: response.data,
                     rawResponse: response.rawResponse,
-                    httpCode: response.httpCode))));
+                    httpCode: response.httpCode)))!);
   }
 
   /// Saves the document first, then submit, in a single db transaction.
@@ -475,22 +479,22 @@ class FrappeModelController extends ModelController<FrappeDocument> {
   ///
   /// Throws [NotSubmittableDocError] if the doctype is not submittable based on its meta.
   @override
-  Future<RequestResponse<T>> saveSubmitDoc<T extends FrappeDocument>(
+  Future<RequestResponse<T?>> saveSubmitDoc<T extends FrappeDocument>(
       T doc) async {
-    await getFrappe().checkAppInstalled(features: ['saveSubmitDoc']);
+    await getFrappe()!.checkAppInstalled(features: ['saveSubmitDoc']);
 
     EmptyDoctypeError.verify(doc.doctype);
     EmptyDocNameError.verify(doc.name);
 
     final meta =
-        await getFrappeMetaController().getDocMeta(doctype: doc.doctype);
+        await getFrappeMetaController()!.getDocMeta(doctype: doc.doctype);
 
-    if (meta.isSuccess && !meta.data.isSubmittable) {
+    if (meta.isSuccess && !meta.data!.isSubmittable!) {
       throw NotSubmittableDocError();
     }
 
     final response = await Request.initiateRequest(
-        url: config.hostUrl,
+        url: config!.hostUrl,
         method: HttpMethod.POST,
         contentType: ContentTypeLiterals.APPLICATION_JSON,
         data: <String, dynamic>{
@@ -499,19 +503,20 @@ class FrappeModelController extends ModelController<FrappeDocument> {
         });
     if (response.isSuccess) {
       if (response.data != null &&
-          response.data.exc == null &&
-          response.data.message is Map) {
-        doc = doc.fromJson(response.data.message);
+          response.data!.exc == null &&
+          response.data!.message is Map) {
+        doc = doc.fromJson(response.data!.message);
         doc.isLocal = false;
         doc.unsaved = false;
-        doc.rawResponse = response.data.message;
+        doc.rawResponse = response.data!.message;
 
         addToLocals(doc);
         return RequestResponse.success(doc, rawResponse: response.rawResponse);
       }
     }
     response.isSuccess = false;
-    return RequestResponse.fail(handleError('save_submit_doc', response.error));
+    return RequestResponse.fail(
+        handleError('save_submit_doc', response.error)!);
   }
 
   /// Returns the cancelled document [T] of a submitted document.
@@ -520,7 +525,7 @@ class FrappeModelController extends ModelController<FrappeDocument> {
   ///
   /// Returns a failure, if the document is duplicated or doctype/docname don't exist.
   @override
-  Future<RequestResponse<T>> cancelDoc<T extends FrappeDocument>(T doc) async {
+  Future<RequestResponse<T?>> cancelDoc<T extends FrappeDocument>(T doc) async {
     EmptyDoctypeError.verify(doc.doctype);
     EmptyDocNameError.verify(doc.name);
 
@@ -528,7 +533,7 @@ class FrappeModelController extends ModelController<FrappeDocument> {
       throw NotASubmittedDocument();
     }
     final response = await Request.initiateRequest(
-        url: config.hostUrl,
+        url: config!.hostUrl,
         method: HttpMethod.POST,
         contentType: ContentTypeLiterals.APPLICATION_JSON,
         data: <String, dynamic>{
@@ -538,19 +543,19 @@ class FrappeModelController extends ModelController<FrappeDocument> {
         });
     if (response.isSuccess) {
       if (response.data != null &&
-          response.data.exc == null &&
-          response.data.message is Map) {
-        doc = doc.fromJson(response.data.message);
+          response.data!.exc == null &&
+          response.data!.message is Map) {
+        doc = doc.fromJson(response.data!.message);
         doc.isLocal = false;
         doc.unsaved = false;
-        doc.rawResponse = response.data.message;
+        doc.rawResponse = response.data!.message;
 
         addToLocals(doc);
         return RequestResponse.success(doc, rawResponse: response.rawResponse);
       }
     }
     response.isSuccess = false;
-    return RequestResponse.fail(handleError('cancel_doc', response.error));
+    return RequestResponse.fail(handleError('cancel_doc', response.error)!);
   }
 
   /// Returns a list of [SearchLinkResponse] after searching against fields.
@@ -569,15 +574,15 @@ class FrappeModelController extends ModelController<FrappeDocument> {
   ///
   /// If the doctype doesn't exist, a failure is returned.
   @override
-  Future<RequestResponse<List<SearchLinkResponse>>> searchLink(
-      {@required String doctype,
-      @required String txt,
-      Map<String, dynamic> options}) async {
+  Future<RequestResponse<List<SearchLinkResponse>?>> searchLink(
+      {required String doctype,
+      required String txt,
+      Map<String, dynamic>? options}) async {
     EmptyDoctypeError.verify(doctype);
 
     options ??= <String, dynamic>{};
     final response = await Request.initiateRequest(
-        url: config.hostUrl,
+        url: config!.hostUrl,
         method: HttpMethod.GET,
         contentType: ContentTypeLiterals.QUERY_STRING,
         queryParams: <String, dynamic>{
@@ -589,15 +594,15 @@ class FrappeModelController extends ModelController<FrappeDocument> {
         isFrappeResponse: false);
     if (response.isSuccess &&
         response.data != null &&
-        response.data.message is Map &&
-        response.data.message['results'] is List) {
-      List temp = response.data.message['results'];
+        response.data!.message is Map &&
+        response.data!.message['results'] is List) {
+      List? temp = response.data!.message['results'];
       final results =
           SearchLinkResponse().deserializeList<SearchLinkResponse>(temp);
       return RequestResponse.success(results,
           rawResponse: response.rawResponse);
     } else {
-      return RequestResponse.fail(handleError('search_link', response.error));
+      return RequestResponse.fail(handleError('search_link', response.error)!);
     }
   }
 
@@ -609,16 +614,16 @@ class FrappeModelController extends ModelController<FrappeDocument> {
   ///
   /// Returns failure in case the [doctype] does not exist in the backend.
   @override
-  Future<RequestResponse<bool>> assignDoc(
-      {@required String assignTo,
-      @required String doctype,
-      bool myself = false,
-      DateTime dueDate,
-      String description,
-      String docName,
-      List<String> docNames,
+  Future<RequestResponse<bool?>> assignDoc(
+      {required String? assignTo,
+      required String doctype,
+      bool? myself = false,
+      DateTime? dueDate,
+      String? description,
+      String? docName,
+      List<String?>? docNames,
       bool notify = false,
-      Priority priority,
+      Priority? priority,
       bool bulkAssign = false}) async {
     // FIXME: possible to assign same doc twice, should be validated
     final args = AssignDocParams(
@@ -633,33 +638,33 @@ class FrappeModelController extends ModelController<FrappeDocument> {
         bulkAssign: bulkAssign,
         notify: notify);
 
-    args.cmd = args.bulkAssign
+    args.cmd = args.bulkAssign!
         ? 'frappe.desk.form.assign_to.add_multiple'
         : 'frappe.desk.form.assign_to.add';
 
-    if (args.myself) args.assignTo = config.coreInstance.auth.currentUser;
+    if (args.myself!) args.assignTo = config!.coreInstance.auth!.currentUser;
 
     final response = await Request.initiateRequest(
-        url: config.hostUrl,
+        url: config!.hostUrl,
         method: HttpMethod.POST,
         contentType: ContentTypeLiterals.APPLICATION_JSON,
         data: (args
               ..name =
-                  args.bulkAssign ? jsonEncode(args.docNames) : args.docName)
+                  args.bulkAssign! ? jsonEncode(args.docNames) : args.docName)
             .toJson());
     if (response.isSuccess) {
       return RequestResponse.success(response.isSuccess);
     } else {
-      return RequestResponse.fail(handleError('assign_doc', response.error));
+      return RequestResponse.fail(handleError('assign_doc', response.error)!);
     }
   }
 
   /// Sets the assignment to [Status.Cancelled] which means that the task is assigned.
   @override
-  Future<RequestResponse<bool>> completeDocAssignment(
-      {@required String doctype,
-      @required String docName,
-      @required String assignedTo}) async {
+  Future<RequestResponse<bool?>> completeDocAssignment(
+      {required String doctype,
+      required String? docName,
+      required String? assignedTo}) async {
     EmptyDoctypeError.verify(doctype);
     EmptyDocNameError.verify(docName);
     final args = CompleteDocAssignmentParams(
@@ -669,7 +674,7 @@ class FrappeModelController extends ModelController<FrappeDocument> {
     )..cmd = 'frappe.desk.form.assign_to.remove';
 
     final response = await Request.initiateRequest(
-        url: config.hostUrl,
+        url: config!.hostUrl,
         method: HttpMethod.POST,
         contentType: ContentTypeLiterals.APPLICATION_JSON,
         data: args.toJson());
@@ -677,23 +682,23 @@ class FrappeModelController extends ModelController<FrappeDocument> {
       return RequestResponse.success(true, rawResponse: response.rawResponse);
     } else {
       return RequestResponse.fail(
-          handleError('complete_doc_assignment', response.error));
+          handleError('complete_doc_assignment', response.error)!);
     }
   }
 
   /// Un-assigns a user from the [doctype] and [docName].
   @override
-  Future<RequestResponse<bool>> unAssignDoc(
-      {@required String doctype,
-      @required String docName,
-      @required String unAssignFrom}) async {
-    await getFrappe().checkAppInstalled(features: ['unAssignDoc']);
+  Future<RequestResponse<bool?>> unAssignDoc(
+      {required String doctype,
+      required String? docName,
+      required String? unAssignFrom}) async {
+    await getFrappe()!.checkAppInstalled(features: ['unAssignDoc']);
 
     EmptyDoctypeError.verify(doctype);
     EmptyDocNameError.verify(docName);
     // FIXME possible to unassign same doc twice, should be validated
     final response = await Request.initiateRequest(
-        url: config.hostUrl,
+        url: config!.hostUrl,
         method: HttpMethod.POST,
         contentType: ContentTypeLiterals.APPLICATION_JSON,
         data: <String, dynamic>{
@@ -705,7 +710,7 @@ class FrappeModelController extends ModelController<FrappeDocument> {
     if (response.isSuccess) {
       return RequestResponse.success(true, rawResponse: response.rawResponse);
     } else {
-      return RequestResponse.fail(handleError('unassign_doc', response.error));
+      return RequestResponse.fail(handleError('unassign_doc', response.error)!);
     }
   }
 
@@ -713,22 +718,24 @@ class FrappeModelController extends ModelController<FrappeDocument> {
   ///
   /// Optionally can filter the list by specifying [doctype] and [status].
   @override
-  Future<RequestResponse<List<GetDocsAssignedToUserResponse>>>
+  Future<RequestResponse<List<GetDocsAssignedToUserResponse>?>>
       getDocsAssignedToUser(
-          {@required String assignedTo, String doctype, Status status}) async {
-    await getFrappe().checkAppInstalled(features: ['getDocsAssignedToUser']);
+          {required String? assignedTo,
+          String? doctype,
+          Status? status}) async {
+    await getFrappe()!.checkAppInstalled(features: ['getDocsAssignedToUser']);
 
     final args = GetDocsAssignedToUserParams(
         assignedTo: assignedTo, doctype: doctype, status: status)
       ..cmd = 'renovation_core.utils.assign_doc.getDocsAssignedToUser';
 
     final response = await Request.initiateRequest(
-        url: config.hostUrl,
+        url: config!.hostUrl,
         method: HttpMethod.POST,
         contentType: ContentTypeLiterals.APPLICATION_JSON,
         data: args.toJson());
     if (response.isSuccess) {
-      List result = response.data.message;
+      List result = response.data!.message;
 
       return RequestResponse.success(
           result.isNotEmpty
@@ -737,14 +744,14 @@ class FrappeModelController extends ModelController<FrappeDocument> {
           rawResponse: response.rawResponse);
     } else {
       return RequestResponse.fail(
-          handleError('get_docs_assigned_to_user', response.error));
+          handleError('get_docs_assigned_to_user', response.error)!);
     }
   }
 
   /// Returns the users assigned to a document of a certain doctype as list of [ToDo].
   @override
-  Future<RequestResponse<List<ToDo>>> getUsersAssignedToDoc(
-      {@required String doctype, @required String docName}) async {
+  Future<RequestResponse<List<ToDo>?>> getUsersAssignedToDoc(
+      {required String doctype, required String docName}) async {
     EmptyDoctypeError.verify(doctype);
     EmptyDocNameError.verify(docName);
     final response = await getList(ToDo(), filters: {
@@ -764,7 +771,7 @@ class FrappeModelController extends ModelController<FrappeDocument> {
       return response;
     } else {
       return RequestResponse.fail(
-          handleError('get_users_assigned_to_doc', response.error));
+          handleError('get_users_assigned_to_doc', response.error)!);
     }
   }
 
@@ -774,14 +781,14 @@ class FrappeModelController extends ModelController<FrappeDocument> {
   ///
   /// Returns failure if the doctype or docname are invalid.
   @override
-  Future<RequestResponse<String>> addTag(
-      {@required String doctype,
-      @required String docName,
-      @required String tag}) async {
+  Future<RequestResponse<String?>> addTag(
+      {required String doctype,
+      required String? docName,
+      required String tag}) async {
     EmptyDoctypeError.verify(doctype);
     EmptyDocNameError.verify(docName);
     final args = {
-      'cmd': config.coreInstance.frappe.frappeVersion.major == 12
+      'cmd': config!.coreInstance.frappe!.frappeVersion!.major == 12
           ? 'frappe.desk.doctype.tag.tag.add_tag'
           : 'frappe.desk.tags.add_tag',
       'dt': doctype,
@@ -790,16 +797,16 @@ class FrappeModelController extends ModelController<FrappeDocument> {
     };
 
     final response = await Request.initiateRequest(
-        url: config.hostUrl,
+        url: config!.hostUrl,
         method: HttpMethod.POST,
         contentType: ContentTypeLiterals.APPLICATION_JSON,
         data: args);
 
     if (response.isSuccess) {
-      return RequestResponse.success(response.data.message,
+      return RequestResponse.success(response.data!.message,
           rawResponse: response.rawResponse);
     } else {
-      return RequestResponse.fail(handleError('add_tag', response.error));
+      return RequestResponse.fail(handleError('add_tag', response.error)!);
     }
   }
 
@@ -809,15 +816,15 @@ class FrappeModelController extends ModelController<FrappeDocument> {
   ///
   /// Returns failure if the doctype or docname are invalid.
   @override
-  Future<RequestResponse<String>> removeTag(
-      {@required String doctype,
-      @required String docName,
-      @required String tag}) async {
+  Future<RequestResponse<String?>> removeTag(
+      {required String doctype,
+      required String? docName,
+      required String tag}) async {
     EmptyDoctypeError.verify(doctype);
     EmptyDocNameError.verify(docName);
 
     final args = {
-      'cmd': config.coreInstance.frappe.frappeVersion.major == 12
+      'cmd': config!.coreInstance.frappe!.frappeVersion!.major == 12
           ? 'frappe.desk.doctype.tag.tag.remove_tag'
           : 'frappe.desk.tags.remove_tag',
       'dt': doctype,
@@ -826,7 +833,7 @@ class FrappeModelController extends ModelController<FrappeDocument> {
     };
 
     final response = await Request.initiateRequest(
-        url: config.hostUrl,
+        url: config!.hostUrl,
         method: HttpMethod.POST,
         contentType: ContentTypeLiterals.APPLICATION_JSON,
         data: args);
@@ -834,7 +841,7 @@ class FrappeModelController extends ModelController<FrappeDocument> {
     if (response.isSuccess) {
       return RequestResponse.success(tag, rawResponse: response.rawResponse);
     } else {
-      return RequestResponse.fail(handleError('remove_tag', response.error));
+      return RequestResponse.fail(handleError('remove_tag', response.error)!);
     }
   }
 
@@ -844,16 +851,16 @@ class FrappeModelController extends ModelController<FrappeDocument> {
   ///
   /// Returns failure if the doctype.
   @override
-  Future<RequestResponse<List<String>>> getTaggedDocs(
-      {@required String doctype, String tag}) async {
+  Future<RequestResponse<List<String>?>> getTaggedDocs(
+      {required String doctype, String? tag}) async {
     EmptyDoctypeError.verify(doctype);
 
     final response = await Request.initiateRequest(
-        url: config.hostUrl,
+        url: config!.hostUrl,
         method: HttpMethod.POST,
         contentType: ContentTypeLiterals.APPLICATION_JSON,
         data: <String, dynamic>{
-          'cmd': config.coreInstance.frappe.frappeVersion.major == 12
+          'cmd': config!.coreInstance.frappe!.frappeVersion!.major == 12
               ? 'frappe.desk.doctype.tag.tag.get_tagged_docs'
               : 'frappe.desk.tags.get_tagged_docs',
           'doctype': doctype,
@@ -861,7 +868,7 @@ class FrappeModelController extends ModelController<FrappeDocument> {
         });
 
     if (response.isSuccess) {
-      List result = response.data.message;
+      List result = response.data!.message;
       if (result.isNotEmpty) {
         return RequestResponse.success(List<String>.from(result[0]),
             rawResponse: response.rawResponse);
@@ -869,7 +876,7 @@ class FrappeModelController extends ModelController<FrappeDocument> {
       return RequestResponse.success([], rawResponse: response.rawResponse);
     } else {
       return RequestResponse.fail(
-          handleError('get_tagged_docs', response.error));
+          handleError('get_tagged_docs', response.error)!);
     }
   }
 
@@ -879,11 +886,11 @@ class FrappeModelController extends ModelController<FrappeDocument> {
   ///
   /// Returns failure if the doctype.
   @override
-  Future<RequestResponse<List<String>>> getTags(
-      {@required String doctype, String likeTag}) async {
-    if (config.coreInstance.frappe.frappeVersion.major < 12) {
+  Future<RequestResponse<List<String?>?>> getTags(
+      {required String doctype, String? likeTag}) async {
+    if (config!.coreInstance.frappe!.frappeVersion!.major! < 12) {
       final response = await Request.initiateRequest(
-          url: config.hostUrl,
+          url: config!.hostUrl,
           method: HttpMethod.POST,
           contentType: ContentTypeLiterals.APPLICATION_JSON,
           data: <String, dynamic>{
@@ -894,11 +901,11 @@ class FrappeModelController extends ModelController<FrappeDocument> {
           });
 
       if (response.isSuccess) {
-        List result = response.data.message;
+        List result = response.data!.message;
         return RequestResponse.success(List<String>.from(result),
             rawResponse: response.rawResponse);
       } else {
-        return RequestResponse.fail(handleError('get_tags', response.error));
+        return RequestResponse.fail(handleError('get_tags', response.error)!);
       }
     } else {
       final response = await getList(TagLink(), filters: [
@@ -909,10 +916,10 @@ class FrappeModelController extends ModelController<FrappeDocument> {
       ]);
 
       if (response.isSuccess) {
-        final tags = response.data.map((_tag) => _tag.tag).toList();
+        final tags = response.data!.map((_tag) => _tag.tag).toList();
         return RequestResponse.success(tags, rawResponse: response.rawResponse);
       }
-      return RequestResponse.fail(handleError('get_tag', response.error));
+      return RequestResponse.fail(handleError('get_tag', response.error)!);
     }
   }
 
@@ -922,16 +929,16 @@ class FrappeModelController extends ModelController<FrappeDocument> {
   ///
   /// Optionally specify the target user. Defaults to `null`.
   @override
-  Future<RequestResponse<FrappeReport>> getReport(
-      {@required String report, dynamic filters, String user}) async {
-    await getFrappe().checkAppInstalled(features: ['getReport']);
+  Future<RequestResponse<FrappeReport?>> getReport(
+      {required String report, dynamic filters, String? user}) async {
+    await getFrappe()!.checkAppInstalled(features: ['getReport']);
 
     if (filters != null) {
       if (!DBFilter.isDBFilter(filters)) throw InvalidFrappeFilter();
     }
     filters ??= <String, dynamic>{};
     final response = await Request.initiateRequest(
-      url: '${config.hostUrl}/api/method/renovation/report',
+      url: '${config!.hostUrl}/api/method/renovation/report',
       method: HttpMethod.POST,
       contentType: ContentTypeLiterals.APPLICATION_JSON,
       data: <String, dynamic>{
@@ -943,28 +950,28 @@ class FrappeModelController extends ModelController<FrappeDocument> {
     );
     if (response.isSuccess && response.data != null) {
       return RequestResponse.success(
-          FrappeReport.fromJson(response.data.message),
+          FrappeReport.fromJson(response.data!.message),
           rawResponse: response.rawResponse);
     } else {
-      return RequestResponse.fail(handleError('get_report', response.error));
+      return RequestResponse.fail(handleError('get_report', response.error)!);
     }
   }
 
   @override
-  ErrorDetail handleError(String errorId, ErrorDetail error) {
+  ErrorDetail? handleError(String? errorId, ErrorDetail? error) {
     switch (errorId) {
       case 'get_doc':
-        if (error.info != null) {
-          if (error.info.httpCode == 404) {
+        if (error!.info != null) {
+          if (error.info!.httpCode == 404) {
             error = handleError('non_existing_doc', error);
-          } else if (error.info.data != null &&
+          } else if (error.info!.data != null &&
               error.info?.data?.exception != null) {
-            if (error.info.data.exception.contains('DoesNotExistError')) {
+            if (error.info!.data.exception.contains('DoesNotExistError')) {
               error = handleError('non_existing_doc', error);
-            } else if (error.info.data.exception.contains('ImportError')) {
+            } else if (error.info!.data.exception.contains('ImportError')) {
               error = handleError('non_existing_doctype', error);
             }
-          } else if (error.info.httpCode == 412) {
+          } else if (error.info!.httpCode == 412) {
             error = handleError('wrong_input', error);
           } else {
             error = handleError(null, error);
@@ -973,26 +980,26 @@ class FrappeModelController extends ModelController<FrappeDocument> {
         break;
       case 'wrong_input':
         error
-          ..type = RenovationError.DataFormatError
+          ?..type = RenovationError.DataFormatError
           ..title = 'Wrong input'
           ..info = (error.info
-            ..httpCode = 412
+            ?..httpCode = 412
             ..cause = 'The input arguments are in the wrong type/format'
             ..suggestion =
                 'Use the correct parameters types/formats referencing the functions signature'
             ..data = error.info);
         break;
       case 'delete_doc':
-        if (error.info.httpCode == 404) {
+        if (error!.info!.httpCode == 404) {
           error = handleError('non_existing', error);
         } else {
           error = handleError(null, error);
         }
         break;
       case 'get_report':
-        if (error.info.httpCode == 404 &&
-            error.info.data != null &&
-            error.info.data.excType.contains('DoesNotExistError')) {
+        if (error!.info!.httpCode == 404 &&
+            error.info!.data != null &&
+            error.info!.data.excType.contains('DoesNotExistError')) {
           error = handleError('non_existing_doctype', error);
         } else {
           error = handleError(null, error);
@@ -1001,28 +1008,28 @@ class FrappeModelController extends ModelController<FrappeDocument> {
       case 'set_value':
       case 'add_tag':
       case 'remove_tag':
-        if (error.info.httpCode == 500) {
+        if (error!.info!.httpCode == 500) {
           error = handleError('non_existing_doctype', error);
-        } else if (error.info.httpCode == 404) {
+        } else if (error.info!.httpCode == 404) {
           error = handleError('non_existing_doc', error);
         } else {
           error = handleError(null, error);
         }
         break;
       case 'get_value':
-        if (error.info.httpCode == 200) {
+        if (error!.info!.httpCode == 200) {
           error = handleError('non_existing', error);
-        } else if (error.info.httpCode == 500) {
+        } else if (error.info!.httpCode == 500) {
           error = handleError('non_existing_docfield', error);
         } else {
           handleError(null, error);
         }
         break;
       case 'save_doc':
-        if (error.info != null &&
-            error.info.data != null &&
-            error.info.data.exception != null &&
-            error.info.data.exception.contains('DuplicateEntryError')) {
+        if (error!.info != null &&
+            error.info!.data != null &&
+            error.info!.data.exception != null &&
+            error.info!.data.exception.contains('DuplicateEntryError')) {
           error = handleError('duplicate_document', error);
         } else {
           error = handleError(null, error);
@@ -1031,29 +1038,29 @@ class FrappeModelController extends ModelController<FrappeDocument> {
       case 'submit_doc':
       case 'save_submit_doc':
       case 'cancel_doc':
-        if (error.info.httpCode == 409) {
+        if (error!.info!.httpCode == 409) {
           error = handleError('duplicate_document', error);
-        } else if (error.info.httpCode == 500) {
+        } else if (error.info!.httpCode == 500) {
           error = handleError('non_existing_doctype', error);
-        } else if (error.info.httpCode == 404) {
+        } else if (error.info!.httpCode == 404) {
           error = handleError('non_existing_doc', error);
         } else {
           error = handleError(null, error);
         }
         break;
       case 'search_link':
-        if (error.info.httpCode == 404 &&
-            error.info.data != null &&
-            error.info.data.excType == 'DoesNotExistError') {
+        if (error!.info!.httpCode == 404 &&
+            error.info!.data != null &&
+            error.info!.data.excType == 'DoesNotExistError') {
           error = handleError('non_existing_doctype', error);
         } else {
           error = handleError(null, error);
         }
         break;
       case 'get_tagged_docs':
-        if (error.info.httpCode == 500 &&
-            error.info.data != null &&
-            error.info.data.exc.contains("doesn't exist")) {
+        if (error!.info!.httpCode == 500 &&
+            error.info!.data != null &&
+            error.info!.data.exc.contains("doesn't exist")) {
           error = handleError('non_existing_doctype', error);
         } else {
           error = handleError(null, error);
@@ -1061,7 +1068,7 @@ class FrappeModelController extends ModelController<FrappeDocument> {
         break;
 
       case 'get_tags':
-        if (error.info.httpCode == 500) {
+        if (error!.info!.httpCode == 500) {
           error = handleError('non_existing_doctype', error);
         } else {
           error = handleError(null, error);
@@ -1069,20 +1076,20 @@ class FrappeModelController extends ModelController<FrappeDocument> {
         break;
       case 'non_existing':
         error
-          ..type = RenovationError.NotFoundError
+          ?..type = RenovationError.NotFoundError
           ..title = DOCTYPE_OR_DOCNAME_NOT_EXIST_TITLE
           ..info = (error.info
-            ..httpCode = 404
+            ?..httpCode = 404
             ..cause = 'Doctype or Docname does not exist'
             ..suggestion = 'Make sure the queried document name is correct'
             ..data = error.info);
         break;
       case 'non_existing_doc':
         error
-          ..type = RenovationError.NotFoundError
+          ?..type = RenovationError.NotFoundError
           ..title = DOCNAME_NOT_EXIST_TITLE
           ..info = (error.info
-            ..httpCode = 404
+            ?..httpCode = 404
             ..cause = 'Docname does not exist'
             ..suggestion =
                 'Make sure the queried document name is correct or create the required document'
@@ -1090,10 +1097,10 @@ class FrappeModelController extends ModelController<FrappeDocument> {
         break;
       case 'non_existing_doctype':
         error
-          ..type = RenovationError.NotFoundError
+          ?..type = RenovationError.NotFoundError
           ..title = DOCTYPE_NOT_EXIST_TITLE
           ..info = (error.info
-            ..httpCode = 404
+            ?..httpCode = 404
             ..cause = 'DocType does not exist'
             ..suggestion =
                 'Make sure the queried DocType is input correctly or create the required DocType'
@@ -1101,10 +1108,10 @@ class FrappeModelController extends ModelController<FrappeDocument> {
         break;
       case 'non_existing_docfield':
         error
-          ..title = 'DocField is not valid'
+          ?..title = 'DocField is not valid'
           ..type = RenovationError.NotFoundError
-          ..info = (error.info
-            ..httpCode = 404
+          ..info = (error!.info
+            ?..httpCode = 404
             ..cause = 'DocField is not defined for the DocType'
             ..suggestion =
                 'Make sure the docfield is input. It is case-sensitive'
@@ -1112,19 +1119,19 @@ class FrappeModelController extends ModelController<FrappeDocument> {
         break;
       case 'duplicate_document':
         error
-          ..title = 'Duplicate document found'
+          ?..title = 'Duplicate document found'
           ..type = RenovationError.DuplicateEntryError
-          ..info = (error.info
-            ..httpCode = 409
+          ..info = (error!.info
+            ?..httpCode = 409
             ..cause = 'Duplicate doc found'
             ..suggestion =
                 'Change the "name" field or delete the existing document'
             ..data = error.info);
         break;
       case 'get_list':
-        if (error.info.httpCode == 500 &&
-            error.info.data != null &&
-            error.info.data.exc.contains('TableMissingError')) {
+        if (error!.info!.httpCode == 500 &&
+            error.info!.data != null &&
+            error.info!.data.exc.contains('TableMissingError')) {
           error = handleError('non_existing_doctype', error);
         }
         break;
@@ -1134,7 +1141,7 @@ class FrappeModelController extends ModelController<FrappeDocument> {
       case 'get_docs_assigned_to_user':
       case 'get_users_assigned_to_doc':
       default:
-        error = RenovationController.genericError(error);
+        error = RenovationController.genericError(error!);
     }
     return error;
   }
