@@ -20,52 +20,52 @@ class GetListParams extends JSONAble with FrappeAPI {
       this.tableFieldsFrappe,
       this.withLinkFieldsFrappe});
 
-  factory GetListParams.fromJson(Map<String, dynamic> json) =>
-      _$GetListParamsFromJson(json);
+  factory GetListParams.fromJson(Map<String, dynamic>? json) =>
+      _$GetListParamsFromJson(json!);
 
-  String doctype;
-  String fields;
+  String? doctype;
+  String? fields;
 
   @JsonKey(name: 'order_by')
-  String orderBy;
+  String? orderBy;
 
   @JsonKey(name: 'limit_start')
-  int limitPageStart;
+  int? limitPageStart;
 
   @JsonKey(name: 'limit_page_length')
-  int limitPageLength;
+  int? limitPageLength;
 
   dynamic filters;
-  String parent;
+  String? parent;
 
   @JsonKey(name: 'table_fields')
-  String tableFieldsFrappe;
+  String? tableFieldsFrappe;
 
   @JsonKey(name: 'with_link_fields')
-  String withLinkFieldsFrappe;
+  String? withLinkFieldsFrappe;
 
   @override
   Map<String, dynamic> toJson() => _$GetListParamsToJson(this);
 
   @override
-  T fromJson<T>(Map<String, dynamic> json) => GetListParams.fromJson(json) as T;
+  T fromJson<T>(Map<String, dynamic>? json) => GetListParams.fromJson(json) as T;
 }
 
 @JsonSerializable()
 class GetValueParams extends JSONAble with FrappeAPI {
   GetValueParams(this.doctype, this.docname, this.docfield);
 
-  factory GetValueParams.fromJson(Map<String, dynamic> json) =>
-      _$GetValueParamsFromJson(json);
+  factory GetValueParams.fromJson(Map<String, dynamic>? json) =>
+      _$GetValueParamsFromJson(json!);
 
-  String doctype;
+  String? doctype;
   @JsonKey(name: 'filters')
-  String docname;
+  String? docname;
   @JsonKey(name: 'fieldname')
-  String docfield;
+  String? docfield;
 
   @override
-  T fromJson<T>(Map<String, dynamic> json) =>
+  T fromJson<T>(Map<String, dynamic>? json) =>
       GetValueParams.fromJson(json) as T;
 
   @override
@@ -76,18 +76,18 @@ class GetValueParams extends JSONAble with FrappeAPI {
 class SetValueParams extends JSONAble with FrappeAPI {
   SetValueParams(this.doctype, this.docname, this.docfield, this.value);
 
-  factory SetValueParams.fromJson(Map<String, dynamic> json) =>
-      _$SetValueParamsFromJson(json);
+  factory SetValueParams.fromJson(Map<String, dynamic>? json) =>
+      _$SetValueParamsFromJson(json!);
 
-  String doctype;
+  String? doctype;
   @JsonKey(name: 'name')
-  String docname;
+  String? docname;
   @JsonKey(name: 'fieldname')
-  String docfield;
+  String? docfield;
   dynamic value;
 
   @override
-  T fromJson<T>(Map<String, dynamic> json) =>
+  T fromJson<T>(Map<String, dynamic>? json) =>
       SetValueParams.fromJson(json) as T;
 
   @override
@@ -98,13 +98,13 @@ class SetValueParams extends JSONAble with FrappeAPI {
 class SearchLinkResponse extends JSONAble {
   SearchLinkResponse();
 
-  factory SearchLinkResponse.fromJson(Map<String, dynamic> json) =>
-      _$SearchLinkResponseFromJson(json);
-  String description;
-  String value;
+  factory SearchLinkResponse.fromJson(Map<String, dynamic>? json) =>
+      _$SearchLinkResponseFromJson(json!);
+  String? description;
+  String? value;
 
   @override
-  T fromJson<T>(Map<String, dynamic> json) =>
+  T fromJson<T>(Map<String, dynamic>? json) =>
       SearchLinkResponse.fromJson(json) as T;
 
   @override
@@ -125,33 +125,33 @@ class AssignDocParams extends JSONAble with FrappeAPI {
       this.priority,
       this.bulkAssign = false});
 
-  factory AssignDocParams.fromJson(Map<String, dynamic> json) =>
-      _$AssignDocParamsFromJson(json);
+  factory AssignDocParams.fromJson(Map<String, dynamic>? json) =>
+      _$AssignDocParamsFromJson(json!);
 
-  String name;
+  String? name;
   @JsonKey(name: 'assign_to')
-  String assignTo;
+  String? assignTo;
   @JsonKey(toJson: FrappeDocFieldConverter.boolToCheck)
-  bool myself;
-  String description;
+  bool? myself;
+  String? description;
   @JsonKey(name: 'date', toJson: FrappeDocFieldConverter.toFrappeDateTime)
-  DateTime dueDate;
-  String doctype;
+  DateTime? dueDate;
+  String? doctype;
   @JsonKey(name: 'docname', ignore: true)
-  String docName;
+  String? docName;
   @JsonKey(name: 'docnames', ignore: true)
-  List<String> docNames;
+  List<String?>? docNames;
   @JsonKey(toJson: FrappeDocFieldConverter.boolToCheck)
-  bool notify;
+  bool? notify;
   @JsonKey()
-  Priority priority;
+  Priority? priority;
 
   /// Use List<String>docnames instead while using bulkAssign
   @JsonKey(name: 'bulk_assign', toJson: FrappeDocFieldConverter.boolToCheck)
-  bool bulkAssign;
+  bool? bulkAssign;
 
   @override
-  T fromJson<T>(Map<String, dynamic> json) =>
+  T fromJson<T>(Map<String, dynamic>? json) =>
       AssignDocParams.fromJson(json) as T;
 
   @override
@@ -161,18 +161,18 @@ class AssignDocParams extends JSONAble with FrappeAPI {
 @JsonSerializable()
 class GetDocsAssignedToUserParams extends JSONAble with FrappeAPI {
   GetDocsAssignedToUserParams(
-      {@required this.assignedTo, this.doctype, this.status});
+      {required this.assignedTo, this.doctype, this.status});
 
-  factory GetDocsAssignedToUserParams.fromJson(Map<String, dynamic> json) =>
-      _$GetDocsAssignedToUserParamsFromJson(json);
+  factory GetDocsAssignedToUserParams.fromJson(Map<String, dynamic>? json) =>
+      _$GetDocsAssignedToUserParamsFromJson(json!);
 
   @JsonKey(name: 'user')
-  String assignedTo;
-  String doctype;
-  Status status;
+  String? assignedTo;
+  String? doctype;
+  Status? status;
 
   @override
-  T fromJson<T>(Map<String, dynamic> json) =>
+  T fromJson<T>(Map<String, dynamic>? json) =>
       GetDocsAssignedToUserParams.fromJson(json) as T;
 
   @override
@@ -193,29 +193,29 @@ class GetDocsAssignedToUserResponse extends JSONAble {
       this.doctype,
       this.docname});
 
-  factory GetDocsAssignedToUserResponse.fromJson(Map<String, dynamic> json) =>
-      _$GetDocsAssignedToUserResponseFromJson(json);
+  factory GetDocsAssignedToUserResponse.fromJson(Map<String, dynamic>? json) =>
+      _$GetDocsAssignedToUserResponseFromJson(json!);
 
   /// Time to finish task (yyyy-mm-dd)
-  DateTime dueDate;
-  Status status;
+  DateTime? dueDate;
+  Status? status;
 
   /// DocName of the User who assigned this doc originally
-  String assignedBy;
+  String? assignedBy;
 
   /// Full name of the assigned by user
-  String assignedByFullName;
+  String? assignedByFullName;
 
   /// The assigned to User
-  String assignedTo;
-  String assignedToFullName;
-  Priority priority;
-  String description;
-  String doctype;
-  String docname;
+  String? assignedTo;
+  String? assignedToFullName;
+  Priority? priority;
+  String? description;
+  String? doctype;
+  String? docname;
 
   @override
-  T fromJson<T>(Map<String, dynamic> json) =>
+  T fromJson<T>(Map<String, dynamic>? json) =>
       GetDocsAssignedToUserResponse.fromJson(json) as T;
 
   @override
@@ -232,18 +232,18 @@ class GetUsersAssignedToDocParams {
 @JsonSerializable()
 class CompleteDocAssignmentParams extends JSONAble with FrappeAPI {
   CompleteDocAssignmentParams(
-      {@required this.doctype, @required this.name, @required this.assignedTo});
+      {required this.doctype, required this.name, required this.assignedTo});
 
-  factory CompleteDocAssignmentParams.fromJson(Map<String, dynamic> json) =>
-      _$CompleteDocAssignmentParamsFromJson(json);
+  factory CompleteDocAssignmentParams.fromJson(Map<String, dynamic>? json) =>
+      _$CompleteDocAssignmentParamsFromJson(json!);
 
-  String doctype;
-  String name;
+  String? doctype;
+  String? name;
   @JsonKey(name: 'assign_to')
-  String assignedTo;
+  String? assignedTo;
 
   @override
-  T fromJson<T>(Map<String, dynamic> json) =>
+  T fromJson<T>(Map<String, dynamic>? json) =>
       CompleteDocAssignmentParams.fromJson(json) as T;
 
   @override
@@ -254,27 +254,27 @@ class CompleteDocAssignmentParams extends JSONAble with FrappeAPI {
 class ToDo extends FrappeDocument {
   ToDo() : super('ToDo');
 
-  factory ToDo.fromJson(Map<String, dynamic> json) => _$ToDoFromJson(json);
+  factory ToDo.fromJson(Map<String, dynamic>? json) => _$ToDoFromJson(json!);
 
-  Status status;
-  Priority priority;
+  Status? status;
+  Priority? priority;
 
-  DateTime date;
-  String description;
+  DateTime? date;
+  String? description;
 
   @JsonKey(name: 'reference_type')
-  String referenceType;
+  String? referenceType;
   @JsonKey(name: 'reference_name')
-  String referenceName;
+  String? referenceName;
 
   @JsonKey(name: 'assigned_by')
-  String assignedBy;
+  String? assignedBy;
 
   @JsonKey(name: 'assigned_by_full_name')
-  String assignedByFullName;
+  String? assignedByFullName;
 
   @override
-  T fromJson<T>(Map<String, dynamic> json) => ToDo.fromJson(json) as T;
+  T fromJson<T>(Map<String, dynamic>? json) => ToDo.fromJson(json) as T;
 
   @override
   Map<String, dynamic> toJson() => _$ToDoToJson(this);
@@ -284,15 +284,15 @@ class ToDo extends FrappeDocument {
 class FrappeReport extends JSONAble {
   FrappeReport();
 
-  factory FrappeReport.fromJson(Map<String, dynamic> json) =>
-      _$FrappeReportFromJson(json);
+  factory FrappeReport.fromJson(Map<String, dynamic>? json) =>
+      _$FrappeReportFromJson(json!);
 
-  List<List<dynamic>> result;
+  List<List<dynamic>>? result;
 
-  List<FrappeReportColumn> columns;
+  List<FrappeReportColumn>? columns;
 
   @override
-  T fromJson<T>(Map<String, dynamic> json) => FrappeReport.fromJson(json) as T;
+  T fromJson<T>(Map<String, dynamic>? json) => FrappeReport.fromJson(json) as T;
 
   @override
   Map<String, dynamic> toJson() => _$FrappeReportToJson(this);
@@ -302,16 +302,16 @@ class FrappeReport extends JSONAble {
 class FrappeReportColumn extends JSONAble {
   FrappeReportColumn();
 
-  factory FrappeReportColumn.fromJson(Map<String, dynamic> json) =>
-      _$FrappeReportColumnFromJson(json);
+  factory FrappeReportColumn.fromJson(Map<String, dynamic>? json) =>
+      _$FrappeReportColumnFromJson(json!);
 
-  String label;
-  String type;
-  String width;
+  String? label;
+  String? type;
+  String? width;
   dynamic options;
 
   @override
-  T fromJson<T>(Map<String, dynamic> json) =>
+  T fromJson<T>(Map<String, dynamic>? json) =>
       FrappeReportColumn.fromJson(json) as T;
 
   @override
@@ -322,26 +322,26 @@ class FrappeReportColumn extends JSONAble {
 class FrappeLog extends FrappeDocument {
   FrappeLog() : super('Renovation Log');
 
-  factory FrappeLog.fromJson(Map<String, dynamic> json) =>
-      _$FrappeLogFromJson(json);
+  factory FrappeLog.fromJson(Map<String, dynamic>? json) =>
+      _$FrappeLogFromJson(json!);
 
-  String title;
+  String? title;
 
-  String content;
+  String? content;
 
-  String type;
+  String? type;
 
-  String request;
+  String? request;
 
-  String response;
+  String? response;
 
-  List<FrappeLogTag> tags;
+  List<FrappeLogTag>? tags;
 
   @JsonKey(name: 'tags_list')
-  List<String> tagsList;
+  List<String>? tagsList;
 
   @override
-  T fromJson<T>(Map<String, dynamic> json) => FrappeLog.fromJson(json) as T;
+  T fromJson<T>(Map<String, dynamic>? json) => FrappeLog.fromJson(json) as T;
 
   @override
   Map<String, dynamic> toJson() => _$FrappeLogToJson(this);
@@ -351,13 +351,13 @@ class FrappeLog extends FrappeDocument {
 class FrappeLogTag extends FrappeDocument {
   FrappeLogTag() : super('Renovation Log Tag Selector');
 
-  factory FrappeLogTag.fromJson(Map<String, dynamic> json) =>
-      _$FrappeLogTagFromJson(json);
+  factory FrappeLogTag.fromJson(Map<String, dynamic>? json) =>
+      _$FrappeLogTagFromJson(json!);
 
-  String tag;
+  String? tag;
 
   @override
-  T fromJson<T>(Map<String, dynamic> json) => FrappeLogTag.fromJson(json) as T;
+  T fromJson<T>(Map<String, dynamic>? json) => FrappeLogTag.fromJson(json) as T;
 
   @override
   Map<String, dynamic> toJson() => _$FrappeLogTagToJson(this);
@@ -367,21 +367,21 @@ class FrappeLogTag extends FrappeDocument {
 class TagLink extends FrappeDocument {
   TagLink() : super('Tag Link');
 
-  factory TagLink.fromJson(Map<String, dynamic> json) =>
-      _$TagLinkFromJson(json);
+  factory TagLink.fromJson(Map<String, dynamic>? json) =>
+      _$TagLinkFromJson(json!);
 
   @JsonKey(name: 'document_type')
-  String documentType;
+  String? documentType;
 
   @JsonKey(name: 'document_name')
-  String documentName;
+  String? documentName;
 
-  String tag;
+  String? tag;
 
-  String title;
+  String? title;
 
   @override
-  T fromJson<T>(Map<String, dynamic> json) => TagLink.fromJson(json) as T;
+  T fromJson<T>(Map<String, dynamic>? json) => TagLink.fromJson(json) as T;
 
   @override
   Map<String, dynamic> toJson() => _$TagLinkToJson(this);
