@@ -17,7 +17,7 @@ import 'interfaces.dart';
 ///
 /// Includes a method for loading basic permissions and checking permission of doctypes/documents.
 class FrappePermissionController extends PermissionController<FrappeDocument> {
-  FrappePermissionController(RenovationConfig config) : super(config);
+  FrappePermissionController(RenovationConfig? config) : super(config);
 
   /// Returns the basic params from the backend for the current user defined by [BasicPermInfo].
   ///
@@ -25,11 +25,11 @@ class FrappePermissionController extends PermissionController<FrappeDocument> {
   ///
   /// If a user isn't signed in, the Guest basic permissions is retrieved.
   @override
-  Future<RequestResponse<BasicPermInfo>> loadBasicPerms() async {
-    await getFrappe().checkAppInstalled(features: ['loadBasicPerms']);
+  Future<RequestResponse<BasicPermInfo?>> loadBasicPerms() async {
+    await getFrappe()!.checkAppInstalled(features: ['loadBasicPerms']);
 
     if (basicPerms != null) {
-      if (basicPerms.isLoading) {
+      if (basicPerms!.isLoading!) {
         await Future<dynamic>.delayed(Duration(milliseconds: 50));
         return loadBasicPerms();
       } else {
@@ -39,16 +39,16 @@ class FrappePermissionController extends PermissionController<FrappeDocument> {
 
     basicPerms = BasicPermInfo()..isLoading = true;
 
-    final user = config.coreInstance.auth.currentUser ?? 'Guest';
+    final user = config!.coreInstance.auth!.currentUser ?? 'Guest';
 
     final response = await Request.initiateRequest(
-        url: config.hostUrl +
+        url: config!.hostUrl! +
             '/api/method/renovation_core.utils.client.get_current_user_permissions',
         method: HttpMethod.POST,
         contentType: ContentTypeLiterals.APPLICATION_JSON);
     if (response.isSuccess && response.data != null) {
-      response.data.message != null
-          ? basicPerms = (BasicPermInfo.fromJson(response.data.message)
+      response.data!.message != null
+          ? basicPerms = (BasicPermInfo.fromJson(response.data!.message)
             ..user = user
             ..isLoading = false)
           : basicPerms = BasicPermInfo();
@@ -63,7 +63,7 @@ class FrappePermissionController extends PermissionController<FrappeDocument> {
   /// Optionally can specify the permission of a certain [document] (TBD).
   @override
   Future<RequestResponse<List<DocPerm>>> getPerm<T extends FrappeDocument>(
-      {@required String doctype, T document}) async {
+      {required String? doctype, T? document}) async {
     var perm = <DocPerm>[
       DocPerm()
         ..read = false
@@ -71,48 +71,48 @@ class FrappePermissionController extends PermissionController<FrappeDocument> {
         ..permLevel = 0
     ];
 
-    if (getFrappeAuthController().currentUser == '<admin_user>') {
+    if (getFrappeAuthController()!.currentUser == '<admin_user>') {
       perm[0].read = true;
     }
 
     final responses = await Future.wait([
-      getFrappeMetaController().getDocMeta(doctype: doctype),
-      getFrappeAuthController().getCurrentUserRoles()
+      getFrappeMetaController()!.getDocMeta(doctype: doctype),
+      getFrappeAuthController()!.getCurrentUserRoles()
     ]);
-    final RequestResponse<DocType> docMeta = responses[0];
-    final RequestResponse<List<String>> currentUserRoles = responses[1];
+    final RequestResponse<DocType?> docMeta = responses[0] as RequestResponse<DocType?>;
+    final RequestResponse<List<String>?>? currentUserRoles = responses[1] as RequestResponse<List<String>?>?;
 
     if (!docMeta.isSuccess) {
       return RequestResponse.success(perm);
     }
 
-    if (!currentUserRoles.isSuccess) {
+    if (!currentUserRoles!.isSuccess) {
       return RequestResponse.success(perm);
     }
 
     // read from meta.permissions
-    for (final doctypePerm in docMeta.data.permissions) {
+    for (final doctypePerm in docMeta.data!.permissions!) {
       // apply only if this DocPerm role is present for currentUser
-      if (!currentUserRoles.data.contains(doctypePerm.role)) {
+      if (!currentUserRoles.data!.contains(doctypePerm.role)) {
         continue;
       }
 
       // permlevels 0,1,2..
       // There's a new perm level, need to grow the list
-      if (perm.length < doctypePerm.permLevel + 1) {
+      if (perm.length < doctypePerm.permLevel! + 1) {
         perm.add(DocPerm());
-        perm[doctypePerm.permLevel].permLevel = doctypePerm.permLevel;
+        perm[doctypePerm.permLevel!].permLevel = doctypePerm.permLevel;
       }
 
       // For User permissions
       // NOTE: this data is required for displaying match rules in ListComponent
       for (final pType in PermissionType.values) {
-        final permission = perm[doctypePerm.permLevel].toJson();
+        final permission = perm[doctypePerm.permLevel!].toJson();
         final permissionType = EnumToString.convertToString(pType);
         permission[permissionType] = permission[permissionType] == 1
             ? permission[permissionType]
             : (doctypePerm.toJson()[permissionType] ?? 0);
-        perm[doctypePerm.permLevel] = DocPerm.fromJson(permission);
+        perm[doctypePerm.permLevel!] = DocPerm.fromJson(permission);
       }
     }
     // TODO : Implement Document Specific rules
@@ -126,10 +126,10 @@ class FrappePermissionController extends PermissionController<FrappeDocument> {
   /// If there are errors, `false` is returned.
   @override
   Future<bool> hasPerm(
-      {@required String doctype,
-      @required PermissionType pType,
-      int permLevel,
-      String docname}) async {
+      {required String? doctype,
+      required PermissionType pType,
+      int? permLevel,
+      String? docname}) async {
     if (!doctypePerms.containsKey(doctype)) {
       final permsR = await getPerm(doctype: doctype);
       if (permsR.isSuccess) {
@@ -152,11 +152,11 @@ class FrappePermissionController extends PermissionController<FrappeDocument> {
     var perm =
         perms[permLevel].toJson()[EnumToString.convertToString(pType)] == 1;
     if (permLevel == 0 && docname != null) {
-      final docInfo = await getFrappeMetaController()
+      final docInfo = await getFrappeMetaController()!
           .getDocInfo(doctype: doctype, docname: docname);
       if (docInfo.isSuccess &&
           docInfo.data != null &&
-          docInfo.data.permissions
+          docInfo.data!.permissions!
                   .toJson()[EnumToString.convertToString(pType)] ==
               0) {
         perm = false;
@@ -170,9 +170,9 @@ class FrappePermissionController extends PermissionController<FrappeDocument> {
   /// If one of the perms is not allowed, the method returns `false`.
   @override
   Future<bool> hasPerms(
-      {@required String doctype,
-      @required List<PermissionType> pTypes,
-      String docname}) async {
+      {required String doctype,
+      required List<PermissionType> pTypes,
+      String? docname}) async {
     for (final pType in pTypes) {
       final hasPerm = await this.hasPerm(
           doctype: doctype, pType: pType, permLevel: 0, docname: docname);
@@ -184,12 +184,12 @@ class FrappePermissionController extends PermissionController<FrappeDocument> {
   }
 
   @override
-  ErrorDetail handleError(String errorId, ErrorDetail error) {
+  ErrorDetail handleError(String errorId, ErrorDetail? error) {
     ErrorDetail err;
     switch (errorId) {
       case 'load_basic_perms':
       default:
-        err = RenovationController.genericError(error);
+        err = RenovationController.genericError(error!);
     }
 
     return err;
