@@ -8,26 +8,26 @@ part 'interfaces.g.dart';
 class BasicPermInfo extends JSONAble {
   BasicPermInfo();
 
-  factory BasicPermInfo.fromJson(Map<String, dynamic> json) =>
-      _$BasicPermInfoFromJson(json);
+  factory BasicPermInfo.fromJson(Map<String, dynamic>? json) =>
+      _$BasicPermInfoFromJson(json!);
 
-  List<String> can_search;
-  List<String> can_email;
-  List<String> can_export;
-  List<String> can_get_report;
-  List<String> can_cancel;
-  List<String> can_print;
-  List<String> can_set_user_permissions;
-  List<String> can_delete;
-  List<String> can_write;
-  List<String> can_import;
-  List<String> can_read;
-  List<String> can_create;
-  bool isLoading;
-  String user;
+  List<String>? can_search;
+  List<String>? can_email;
+  List<String>? can_export;
+  List<String>? can_get_report;
+  List<String>? can_cancel;
+  List<String>? can_print;
+  List<String>? can_set_user_permissions;
+  List<String>? can_delete;
+  List<String>? can_write;
+  List<String>? can_import;
+  List<String>? can_read;
+  List<String>? can_create;
+  bool? isLoading;
+  String? user;
 
   @override
-  T fromJson<T>(Map<String, dynamic> json) => BasicPermInfo.fromJson(json) as T;
+  T fromJson<T>(Map<String, dynamic>? json) => BasicPermInfo.fromJson(json) as T;
 
   @override
   Map<String, dynamic> toJson() => _$BasicPermInfoToJson(this);
