@@ -19,8 +19,8 @@ import 'frappe.socketiouploader.dart';
 import 'interfaces.dart';
 
 class FrappeStorageController extends StorageController<FrappeUploadFileParams,
-    FrappeUploadFileResponse, FrappeUploadStatus> {
-  FrappeStorageController(RenovationConfig config) : super(config);
+    FrappeUploadFileResponse?, FrappeUploadStatus> {
+  FrappeStorageController(RenovationConfig? config) : super(config);
 
   /// Validate the arguments of the file to be uploaded.
   ///
@@ -35,7 +35,7 @@ class FrappeStorageController extends StorageController<FrappeUploadFileParams,
     if (uploadFileParams.file == null) throw MissingFileError();
 
     if (uploadFileParams.fileName == null ||
-        uploadFileParams.fileName.isEmpty) {
+        uploadFileParams.fileName!.isEmpty) {
       throw MissingFileNameError();
     }
 
@@ -62,7 +62,7 @@ class FrappeStorageController extends StorageController<FrappeUploadFileParams,
         case UploadingStatus.ready:
           break;
         case UploadingStatus.error:
-          config.coreInstance.config.logger.w([
+          config!.coreInstance.config!.logger.w([
             'LTS-Renovation-Core',
             'Frappe SocketIO Upload error',
             EnumToString.convertToString(uploadStatus.error)
@@ -81,7 +81,7 @@ class FrappeStorageController extends StorageController<FrappeUploadFileParams,
           await obs.close();
           break;
         case UploadingStatus.uploading:
-          config.coreInstance.config.logger
+          config!.coreInstance.config!.logger
               .i(['Upload Progress', uploadStatus.progress.toString()]);
           obs.add(FrappeUploadStatus()
             ..filename = uploadStatus.filename
@@ -100,11 +100,11 @@ class FrappeStorageController extends StorageController<FrappeUploadFileParams,
           break;
         case UploadingStatus.detail_error:
         default:
-          RequestResponse<FrappeUploadFileResponse> response;
+          RequestResponse<FrappeUploadFileResponse?>? response;
           if (uploadStatus.r != null) {
             response = uploadStatus.r;
           } else {
-            response = RequestResponse.fail(handleError(null, null));
+            response = RequestResponse.fail(handleError(null, null)!);
           }
           obs.add(FrappeUploadStatus()
             ..status = UploadingStatus.error
@@ -123,9 +123,9 @@ class FrappeStorageController extends StorageController<FrappeUploadFileParams,
   ///
   /// Handles files of type [File] if native and [Uint8List] if from the web.
   @override
-  Future<RequestResponse<FrappeUploadFileResponse>> uploadViaHTTP(
+  Future<RequestResponse<FrappeUploadFileResponse?>> uploadViaHTTP(
       FrappeUploadFileParams uploadFileParams) async {
-    await getFrappe().checkAppInstalled(features: ['uploadViaHTTP']);
+    await getFrappe()!.checkAppInstalled(features: ['uploadViaHTTP']);
 
     validateUploadFileArgs(uploadFileParams);
 
@@ -134,18 +134,19 @@ class FrappeStorageController extends StorageController<FrappeUploadFileParams,
 
     uploadFileParams.fileSize = getFileSize(uploadFileParams.file);
 
-    uploadFileParams.cmd = getFrappe().getAppsVersion('renovation_core') != null
-        ? 'renovation_core.handler.uploadfile'
-        : 'frappe.handler.uploadfile';
+    uploadFileParams.cmd =
+        getFrappe()!.getAppsVersion('renovation_core') != null
+            ? 'renovation_core.handler.uploadfile'
+            : 'frappe.handler.uploadfile';
 
     final response = await Request.initiateRequest(
-        url: config.hostUrl,
+        url: config!.hostUrl,
         method: HttpMethod.POST,
         contentType: ContentTypeLiterals.APPLICATION_X_WWW_FORM_URLENCODED,
         data: uploadFileParams.toJson());
 
     if (response.isSuccess) {
-      final dynamic responseObj = response.data.message;
+      final dynamic responseObj = response.data!.message;
       if (responseObj != null) {
         final uploadFileResponse = FrappeUploadFileResponse()
             .fromJson<FrappeUploadFileResponse>(responseObj);
@@ -157,10 +158,11 @@ class FrappeStorageController extends StorageController<FrappeUploadFileParams,
             'uploadViaHttp',
             response.error ??
                 ErrorDetail(
-                    info: Information(rawResponse: response.rawResponse))));
+                    info: Information(rawResponse: response.rawResponse)))!);
       }
     } else {
-      return RequestResponse.fail(handleError('uploadViaHttp', response.error));
+      return RequestResponse.fail(
+          handleError('uploadViaHttp', response.error)!);
     }
   }
 
@@ -168,14 +170,14 @@ class FrappeStorageController extends StorageController<FrappeUploadFileParams,
   ///
   /// It calls `getList()` of [FrappeModelController].
   @override
-  Future<RequestResponse<bool>> checkFolderExists(String folderDir) async {
-    final response = await config.coreInstance.model.getList(FrappeFile(),
+  Future<RequestResponse<bool?>> checkFolderExists(String folderDir) async {
+    final response = await config!.coreInstance.model!.getList(FrappeFile(),
         fields: ['name'], filters: {'is_folder': 1, 'name': folderDir});
     if (response.isSuccess) {
-      return RequestResponse.success(response.data.isNotEmpty);
+      return RequestResponse.success(response.data!.isNotEmpty);
     } else {
       return RequestResponse.fail(
-          handleError('checkFolderExists', response.error));
+          handleError('checkFolderExists', response.error)!);
     }
   }
 
@@ -187,16 +189,16 @@ class FrappeStorageController extends StorageController<FrappeUploadFileParams,
   ///
   /// [folderName] must not contain a forward-slash, otherwise an error is returned within [RequestResponse].
   @override
-  Future<RequestResponse<bool>> createFolder(
-      {@required String folderName, String parentFolder = 'Home'}) async {
+  Future<RequestResponse<bool?>> createFolder(
+      {required String folderName, String? parentFolder = 'Home'}) async {
     if (folderName.contains('/')) {
       return RequestResponse.fail(handleError(
           'invalid_foldername_forward_slash',
-          ErrorDetail(info: Information(httpCode: 412))));
+          ErrorDetail(info: Information(httpCode: 412)))!);
     }
 
     final response = await Request.initiateRequest(
-        url: config.hostUrl,
+        url: config!.hostUrl,
         method: HttpMethod.POST,
         contentType: ContentTypeLiterals.APPLICATION_X_WWW_FORM_URLENCODED,
         data: <String, dynamic>{
@@ -207,34 +209,34 @@ class FrappeStorageController extends StorageController<FrappeUploadFileParams,
 
     return response.isSuccess
         ? RequestResponse.success(true, rawResponse: response.rawResponse)
-        : RequestResponse.fail(handleError('create_folder', response.error));
+        : RequestResponse.fail(handleError('create_folder', response.error)!);
   }
 
   /// Returns the URL appended to the reference of the file using the `hostUrl`.
   ///
   /// In case [ref] is `null` or contains "http", [ref] is returned as-is.
   @override
-  String getUrl(String ref) {
+  String? getUrl(String? ref) {
     if (ref == null) return ref;
 
     // Already a well formed url, return it as-is.
     if (ref.contains('http')) return ref;
 
-    return '${config.hostUrl}$ref';
+    return '${config!.hostUrl}$ref';
   }
 
   @override
-  ErrorDetail handleError(String errorId, ErrorDetail error) {
+  ErrorDetail? handleError(String? errorId, ErrorDetail? error) {
     switch (errorId) {
       case 'create_folder':
-        if (error.info.httpCode == 412) {
+        if (error!.info!.httpCode == 412) {
           error = handleError('invalid_foldername_forward_slash', error);
-        } else if (error.info.httpCode == 409) {
+        } else if (error.info!.httpCode == 409) {
           error
             ..title = 'Folder with same name exists'
             ..type = RenovationError.DuplicateEntryError
             ..info = (error.info
-              ..httpCode = 409
+              ?..httpCode = 409
               ..cause = 'A folder with same name exists'
               ..suggestion =
                   'Choose a new name for the folder or remove the existing one');
@@ -244,20 +246,20 @@ class FrappeStorageController extends StorageController<FrappeUploadFileParams,
         break;
       case 'invalid_foldername_forward_slash':
         error
-          ..title = 'Invalid Folder Name'
+          ?..title = 'Invalid Folder Name'
           ..info = (error.info
-            ..httpCode = 412
+            ?..httpCode = 412
             ..cause = 'Invalid folder name: contains forward slash'
             ..suggestion = 'Remove the forward slash from the name');
         break;
       case 'uploadViaHttp':
-        if (error.info.rawResponse.data
+        if (error!.info!.rawResponse!.data!
             .contains('Same file has already been attached to the record')) {
           error
             ..type = RenovationError.DuplicateEntryError
             ..title = 'File already exists with attached to the document'
             ..info = (error.info
-              ..httpCode = 409
+              ?..httpCode = 409
               ..cause = 'Same file was uploaded earlier to the document'
               ..suggestion =
                   'Either delete the file before re-uploading or ignore the message');
@@ -267,7 +269,7 @@ class FrappeStorageController extends StorageController<FrappeUploadFileParams,
         break;
       case 'checkFolderExists':
       default:
-        error = RenovationController.genericError(error);
+        error = RenovationController.genericError(error!);
     }
 
     return error;
