@@ -6,11 +6,11 @@ import '../core/request.dart';
 
 /// Class handling the translations and setting the language
 abstract class TranslationController extends RenovationController {
-  TranslationController(RenovationConfig config) : super(config);
+  TranslationController(RenovationConfig? config) : super(config);
 
   /// The current set language. Defaults to English.
   @protected
-  String currentLanguage = 'en';
+  String? currentLanguage = 'en';
 
   /// Messages dictionary holding translation maps of each language. Structure is:
   ///```
@@ -27,14 +27,15 @@ abstract class TranslationController extends RenovationController {
       <String, Map<String, String>>{};
 
   /// Loads the translation of the selected language from the backend.
-  Future<RequestResponse<Map<String, String>>> loadTranslations({String lang});
+  Future<RequestResponse<Map<String, String>?>> loadTranslations(
+      {String? lang});
 
   /// Returns the translated text paired with the key [txt].
   ///
   /// If [txt] doesn't exist for a certain language or the [lang] is not defined in the Map,
   ///
   /// [txt] is returned as-is.
-  String getMessage({@required String txt, String lang}) {
+  String? getMessage({required String? txt, String? lang}) {
     lang ??= currentLanguage;
     if (txt == null) {
       return txt;
@@ -47,27 +48,27 @@ abstract class TranslationController extends RenovationController {
   /// Note: This will overwrite the existing map.
   ///
   /// If `null` is passed as [dict], an empty Map will be set instead.
-  void setMessagesDict({@required Map<String, String> dict, String lang}) {
+  void setMessagesDict({required Map<String, String> dict, String? lang}) {
     lang ??= currentLanguage;
-    _messages[lang] = dict ?? <String, String>{};
+    _messages[lang!] = dict ?? <String, String>{};
   }
 
   /// Add translations to the existing Map.
   ///
   /// Unlike [setMessagesDict], this will not overwrite the existing map, just adds to it (or replace in case the key is duplicated).
-  void extendDictionary({@required Map<String, String> dict, String lang}) {
+  void extendDictionary({required Map<String, String>? dict, String? lang}) {
     lang ??= currentLanguage;
     if (dict != null && dict is Map<String, String>) {
-      _messages[lang] ??= <String, String>{};
-      _messages[lang].addAll(dict);
+      _messages[lang!] ??= <String, String>{};
+      _messages[lang]!.addAll(dict);
     }
   }
 
   /// A setter for [currentLanguage]. Defaults to 'en'
-  void setCurrentLanguage(String lang) => currentLanguage = lang;
+  void setCurrentLanguage(String? lang) => currentLanguage = lang;
 
   /// Returns the [currentLanguage].
-  String getCurrentLanguage() => currentLanguage;
+  String? getCurrentLanguage() => currentLanguage;
 
   /// Removes all translations saved in messages dictionary.
   @override
