@@ -8,13 +8,13 @@ part 'models.g.dart';
 class RenovationUserAgreement extends FrappeDocument {
   RenovationUserAgreement() : super('Renovation User Agreement');
 
-  factory RenovationUserAgreement.fromJson(Map<String, dynamic> json) =>
-      _$RenovationUserAgreementFromJson(json);
+  factory RenovationUserAgreement.fromJson(Map<String, dynamic>? json) =>
+      _$RenovationUserAgreementFromJson(json!);
 
-  String title;
+  String? title;
 
   @override
-  T fromJson<T>(Map<String, dynamic> json) =>
+  T fromJson<T>(Map<String, dynamic>? json) =>
       RenovationUserAgreement.fromJson(json) as T;
 
   @override
@@ -25,26 +25,26 @@ class RenovationUserAgreement extends FrappeDocument {
 class RenovationReview extends FrappeDocument {
   RenovationReview() : super('Renovation Review');
 
-  factory RenovationReview.fromJson(Map<String, dynamic> json) =>
-      _$RenovationReviewFromJson(json);
+  factory RenovationReview.fromJson(Map<String, dynamic>? json) =>
+      _$RenovationReviewFromJson(json!);
 
   @JsonKey(name: 'reviewed_by_doctype')
-  String reviewedByDoctype;
+  String? reviewedByDoctype;
   @JsonKey(name: 'reviewed_by')
-  String reviewedBy;
+  String? reviewedBy;
 
   @JsonKey(name: 'reviewed_by_doctype_doc')
-  DocType reviewedByDoctypeDoc;
+  DocType? reviewedByDoctypeDoc;
 
   @JsonKey(name: 'reviewed_doctype')
-  String reviewedDoctype;
+  String? reviewedDoctype;
   @JsonKey(name: 'reviewed_entity')
-  String reviewedEntity;
+  String? reviewedEntity;
 
-  List<RenovationReviewItem> reviews;
+  List<RenovationReviewItem>? reviews;
 
   @override
-  T fromJson<T>(Map<String, dynamic> json) =>
+  T fromJson<T>(Map<String, dynamic>? json) =>
       RenovationReview.fromJson(json) as T;
 
   @override
@@ -55,16 +55,16 @@ class RenovationReview extends FrappeDocument {
 class RenovationReviewItem extends FrappeDocument {
   RenovationReviewItem() : super('Renovation Review Item');
 
-  factory RenovationReviewItem.fromJson(Map<String, dynamic> json) =>
-      _$RenovationReviewItemFromJson(json);
+  factory RenovationReviewItem.fromJson(Map<String, dynamic>? json) =>
+      _$RenovationReviewItemFromJson(json!);
 
-  String title;
-  String question;
-  String quantitative;
-  String answer;
+  String? title;
+  String? question;
+  String? quantitative;
+  String? answer;
 
   @override
-  T fromJson<T>(Map<String, dynamic> json) =>
+  T fromJson<T>(Map<String, dynamic>? json) =>
       RenovationReviewItem.fromJson(json) as T;
 
   @override
@@ -75,11 +75,11 @@ class RenovationReviewItem extends FrappeDocument {
 class NonExistingDocType extends FrappeDocument {
   NonExistingDocType() : super('NON-EXISTING DOCTYPE');
 
-  factory NonExistingDocType.fromJson(Map<String, dynamic> json) =>
-      _$NonExistingDocTypeFromJson(json);
+  factory NonExistingDocType.fromJson(Map<String, dynamic>? json) =>
+      _$NonExistingDocTypeFromJson(json!);
 
   @override
-  T fromJson<T>(Map<String, dynamic> json) =>
+  T fromJson<T>(Map<String, dynamic>? json) =>
       NonExistingDocType.fromJson(json) as T;
 
   @override
