@@ -29,7 +29,7 @@ class EmptyDoctypeError extends Error {
   @override
   String toString() => 'The doctype should not be null or empty';
 
-  static void verify(String doctype) {
+  static void verify(String? doctype) {
     if (doctype == null || doctype.isEmpty) throw EmptyDoctypeError();
   }
 }
@@ -38,7 +38,7 @@ class EmptyDocNameError extends Error {
   @override
   String toString() => 'The docname should not be null or empty';
 
-  static void verify(String docName) {
+  static void verify(String? docName) {
     if (docName == null || docName.isEmpty) throw EmptyDocNameError();
   }
 }
